Arrays

MPIStateArrays

Storage for the state of a discretization.

CLIMA.MPIStateArrays.MPIStateArray — Type

MPIStateArray{FT, DATN<:AbstractArray{FT,3}, DAI1, DAV,
              DAT2<:AbstractArray{FT,2}} <: AbstractArray{FT, 3}

CLIMA.MPIStateArrays.post_Irecvs! — Function

post_Irecvs!(Q::MPIStateArray)

posts the MPI.Irecv! for Q

CLIMA.MPIStateArrays.finish_ghost_recv! — Function

finish_ghost_recv!(Q::MPIStateArray)

Complete the receive of data and fill the data array on the device

CLIMA.MPIStateArrays.finish_ghost_send! — Function

finish_ghost_send!(Q::MPIStateArray)

Waits on the send of data to be complete

CLIMA.MPIStateArrays.start_ghost_exchange! — Function

start_ghost_exchange!(Q::MPIStateArray; dorecvs=true)

Start the MPI exchange of the data stored in Q. If dorecvs is true then post_Irecvs!(Q) is called, otherwise the caller is responsible for this.

This function will fill the send buffer (on the device), copies the data from the device to the host, and then issues the send. Previous sends are waited on to ensure that they are complete.

CLIMA.MPIStateArrays.finish_ghost_exchange! — Function

finish_ghost_exchange!(Q::MPIStateArray)

Complete the exchange of data and fill the data array on the device. Note this completes both the send and the receive communication. For more fine level control see finish_ghost_recv! and finish_ghost_send!