TurbulenceConvection

Grids

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.Grid — Type

Grid{T, I}

A simple 1-dimensional uniform grid of type T for finite difference method.

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.over_elems — Function

over_elems(grid::Grid)

Get the range of indexes to traverse real and ghost grid elements

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.over_elems_real — Function

over_elems_real(grid::Grid)

Get the range of indexes to traverse only real grid elements

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.over_elems_ghost — Function

over_elems_ghost(grid::Grid)

Get the range of indexes to traverse only ghost grid elements

over_elems_ghost(grid::Grid, ::ZBoundary)

Get the range of indexes to traverse only ghost grid elements below $z_{min}$ and above $z_{max}$.

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.Zmin — Type

Zmin

Type to dispatch function to particular location

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.Zmax — Type

Zmax

Type to dispatch function to particular location

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.n_hat — Function

n_hat(::ZBoundary)

The outward normal vector to the boundary

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.binary — Function

binary(::ZBoundary)

Returns 0 for Zmin and 1 for Zmax

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.ghost_vec — Function

ghost_vec(::ZBoundary)

A 3-element vector near the boundary with 1's on ghost cells and 0's on interior cells

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.ghost_dual — Function

ghost_dual(::ZBoundary)

A 2-element vector near the boundary with 1's on ghost cells and 0's on interior cells

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.first_interior — Function

first_interior(grid::Grid, ::ZBoundary)

Get the first element index in the physical domain

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.boundary — Function

boundary(grid::Grid, ::ZBoundary)

Get the element index for edge data at the boundary

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.second_interior — Function

second_interior(grid::Grid, ::ZBoundary)

Get the second element index in the physical domain

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.first_ghost — Function

first_ghost(grid::Grid, ::ZBoundary)

Get the first element index in the ghost domain

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.boundary_points — Function

boundary_points(grid::Grid, ::ZBoundary)

Get the first element index in the ghost domain

Operators

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.grad — Function

grad(f, grid::Grid{FT}) where FT

Computes local finite-difference gradient of f: ∇f

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_pos — Function

∇_pos(f, grid::Grid)

Computes a one-sided (up) local finite-difference gradient of f: ∇f

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_neg — Function

∇_neg(f, grid::Grid)

Computes a one-sided (down) local finite-difference gradient of f: ∇f

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_z_flux — Function

∇_z_flux(f, grid::Grid)

Computes the local derivative of field $f$, which is assumed to live on the dual grid: $∇f$

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_z_centered — Function

∇_z_centered(f, grid::Grid)

Computes the local derivative of field $f$, which is assumed to live on the dual grid: $∇f$

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_z_dual — Function

∇_z_dual(f, grid::Grid)

Computes the local derivative of field $f$, which is assumed to live on the dual grid: $∇f$

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.∇_z_upwind — Function

∇_z_upwind(ϕ, w, grid::Grid)

Local upwind gradient operator $∇ϕ$. This operator is stable but numerically diffusive.

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.Δ_z — Function

Δ_z(f, grid::Grid)

Computes the local Laplacian of field $f$: $∇ • (∇f)$

Δ_z(f, grid::Grid, K)

Computes the local Laplacian of field $f$ with a variable coefficient: $∇ • (K ∇f)$

CLIMA.TurbulenceConvection.FiniteDifferenceGrids.Δ_z_dual — Function

Δ_z_dual(f, grid::Grid, K)

Computes the local Laplacian of field $f$ with a variable coefficient: $∇ • (K ∇f)$

State Vector

CLIMA.TurbulenceConvection.StateVecs.StateVec — Type

StateVec{VM}

A state vector containing the number of subdomains, n_subdomains, a NamedTuple variable mapper, a tuple of the variable names, and a vector of vectors, containing the values for all of the variables.

CLIMA.TurbulenceConvection.StateVecs.over_sub_domains — Function

over_sub_domains(sv::StateVec, ϕ::Symbol)

Get list of indexes over all subdomains for variable ϕ.

CLIMA.TurbulenceConvection.StateVecs.Cut — Type

Cut{I} <: AbstractCut{I}

A Cut struct used to slice the state vector along the grid-element dimension. This is used as an API to pass Cuts into local derivative/interpolation routines.

CLIMA.TurbulenceConvection.StateVecs.Dual — Type

Dual{I} <: AbstractCut{I}

A Dual struct used to slice the state vector along the grid-element dimension. In addition, this interpolates fields [(f[k-1]+f[k])/2, (f[k]+f[k+1])/2]

CLIMA.TurbulenceConvection.StateVecs.var_names — Function

var_names(sv::StateVec)

Get tuple of variable names.

CLIMA.TurbulenceConvection.StateVecs.var_string — Function

var_string(sv::StateVec, ϕ::Symbol, i=0)

A string of the variable ϕ with the appropriate suffix indicating its sub-domain.

CLIMA.TurbulenceConvection.StateVecs.var_suffix — Function

var_suffix(sv::StateVec, ϕ::Symbol, i=0)

A suffix string for variable ϕ indicating its sub-domain.

CLIMA.TurbulenceConvection.StateVecs.assign! — Function

assign!(sv::StateVec, var_names, grid::Grid, val, i=1)

Assign value val to variable ϕ for all ghost points.

assign!(sv::StateVec, grid::Grid, val)

Assign value val to all variables in state vector.

CLIMA.TurbulenceConvection.StateVecs.assign_real! — Function

assign_real!(sv::StateVec, ϕ::Symbol, grid::Grid, vec_real, i=0)

Assign the real elements to the given vector vec_real.

CLIMA.TurbulenceConvection.StateVecs.assign_ghost! — Function

assign_ghost!(sv::StateVec, ϕ::Symbol, grid::Grid, val, i=0)

Assign value val to variable ϕ for all ghost points.

assign_ghost!(sv::StateVec, ϕ::Symbol, grid::Grid, val, b::ZBoundary, i=1)

Assign value val to variable ϕ for all ghost points.

CLIMA.TurbulenceConvection.StateVecs.extrap! — Function

extrap!(sv::StateVec, ϕ::Symbol, grid::Grid, i=1)

Extrapolate variable ϕ to the first ghost point.

extrap!(sv::StateVec, ϕ::Symbol, grid::Grid, ::ZBoundary, i=1)

Extrapolate variable ϕ to the first ghost point.

extrap!(sv::StateVec, ϕ::Symbol, grid::Grid, dual_val, b::ZBoundary, i=1)

Extrapolate variable ϕ to the first ghost point.

CLIMA.TurbulenceConvection.StateVecs.extrap_0th_order! — Function

extrap_0th_order!(sv::StateVec, ϕ::Symbol, grid::Grid, i=1)

Extrapolate variable ϕ to the first ghost point using zeroth order approximation.

CLIMA.TurbulenceConvection.StateVecs.compare — Function

compare(sv::StateVec, sv_expected::StateVec, grid::Grid, tol)

A dictionary, with keys in var_names(vs), containing a Bool indicating that sv ≈ sv_expected for all of their sub-domains for all elements.

CLIMA.TurbulenceConvection.StateVecs.DomainIdx — Type

DomainIdx{GM,EN,UD}

Decomposition of domain indexes, including indexes for

i_gm grid-mean domain

i_en environment sub-domain

i_ud updraft sub-domains

Note that the index ordering logic is defined here.

CLIMA.TurbulenceConvection.StateVecs.subdomains — Function

subdomains(idx::DomainIdx)

Tuple of sub-domain indexes, including environment and updrafts

CLIMA.TurbulenceConvection.StateVecs.alldomains — Function

alldomains(idx::DomainIdx)

Tuple of all sub-domain indexes, including grid-mean, environment and updrafts

CLIMA.TurbulenceConvection.StateVecs.eachdomain — Function

eachdomain(idx::DomainIdx)

Tuple of indexes including

grid-mean

environment

Tuple of updraft indexes

CLIMA.TurbulenceConvection.StateVecs.allcombinations — Function

allcombinations(idx::DomainIdx)

Tuple of indexes including

grid-mean

environment

Tuple of updraft indexes

Tuple of all sub-domains

Tuple of all domains

CLIMA.TurbulenceConvection.StateVecs.DomainSubSet — Type

DomainSubSet{GM,EN,UD}

Subset of domains

gm bool indicating to include grid-mean domain

en bool indicating to include environment sub-domain

ud bool indicating to include updraft sub-domains

CLIMA.TurbulenceConvection.StateVecs.get_param — Function

get_param(::Type{Domain})

The number of domains in domain Domain

CLIMA.TurbulenceConvection.StateVecs.DomainDecomp — Type

DomainDecomp{GM,EN,UD}

Decomposition of number of domains

gm number of grid-mean domain

en number of environment sub-domain

ud number of updraft sub-domains

Examples

Several examples exist in the test directory.