CLIMAAtmosDycore

CLIMAAtmosDycore

CLIMA.CLIMAAtmosDycore.getrhsfunction — Function.

getrhsfunction(disc::AbstractAtmosDiscretization)

The spatial discretizations are of the form $Q̇ = f(Q)$, and this function returns the handle to right-hand side function $f$ of the disc

source

CLIMA.CLIMAAtmosDycore.solve! — Function.

solve!(Q, solver::AbstractAtmosODESolver; timeend,
       stopaftertimeend=true, numberofsteps, callbacks)

Solves an ODE using the solver starting from a state Q. The state Q is updated inplace. The final time timeend or numberofsteps must be specified.

A series of optional callback functions can be specified using the tuple callbacks; see GenericCallbacks.

Todo

Currently stopaftertimeend is not used. The idea behind it was that a user might want to stop either one step before or after the final time timeend. This should either be removed or used.

source

AtmosStateArray

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.AtmosStateArray — Type.

AtmosStateArray{S <: Tuple, T, DeviceArray, N,
                DATN<:AbstractArray{T,N}, Nm1, DAI1} <: AbstractArray{T, N}

N-dimensional MPI-aware array with elements of type T. The dimension N is length(S) + 1. S is a tuple of the first N-1 array dimensions.

Todo

It should be reevaluated whether all this stuff in the type domain is really necessary (some of it was optimistically added for functionality that never panned out)

Todo

tag for the MPI message should probably be unified for each AtmosStateArray (right now 888 used is the same for all communication)

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.postrecvs! — Function.

postrecvs!(Q::AtmosStateArray)

posts the MPI.Irecv! for Q

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.startexchange! — Function.

startexchange!(Q::AtmosStateArray; dorecvs=true)

Start the MPI exchange of the data stored in Q. If dorecvs is true then postrecvs!(Q) is called, otherwise the caller is responsible for this.

This function will fill the send buffer (on the device), copies the data from the device to the host, and then issues the send. Previous sends are waited on to ensure that they are complete.

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.finishexchange! — Function.

finishexchange!(Q::AtmosStateArray)

Complete the exchange of data and fill the data array on the device

source

Grids

CLIMA.CLIMAAtmosDycore.Grids.DiscontinuousSpectralElementGrid — Type.

DiscontinuousSpectralElementGrid(topology; FloatType, DeviceArray,
                                 polynomialorder,
                                 meshwarp = (x...)->identity(x))

Generate a discontinuous spectral element (tensor product, Legendre-Gauss-Lobatto) grid/mesh from a topology, where the order of the elements is given by polynomialorder. DeviceArray gives the array type used to store the data (CuArray or Array), and the coordinate points will be of FloatType.

The optional meshwarp function allows the coordinate points to be warped after the mesh is created; the mesh degrees of freedom are orginally assigned using a trilinear blend of the element corner locations.

source

Topologies

CLIMA.CLIMAAtmosDycore.Topologies.BrickTopology — Type.

BrickTopology{dim, T}(mpicomm, elemrange; boundary, periodicity)

Generate a brick mesh topology with coordinates given by the tuple elemrange and the periodic dimensions given by the periodicity tuple.

The elements of the brick are partitioned equally across the MPI ranks based on a space-filling curve.

By default boundary faces will be marked with a one and other faces with a zero. Specific boundary numbers can also be passed for each face of the brick in boundary. This will mark the nonperiodic brick faces with the given boundary number.

Examples

We can build a 3 by 2 element two-dimensional mesh that is periodic in the $x_2$-direction with

using CLIMAAtmosDycore
using CLIMAAtmosDycore.Topologies
using MPI
MPI.Init()
topology = BrickTopology(MPI.COMM_SELF, (2:5,4:6);
                         periodicity=(false,true),
                         boundary=[1 3; 2 4])
MPI.Finalize()

This returns the mesh structure for

         x_2

          ^
          |
         6-  +-----+-----+-----+
          |  |     |     |     |
          |  |  3  |  4  |  5  |
          |  |     |     |     |
         5-  +-----+-----+-----+
          |  |     |     |     |
          |  |  1  |  2  |  6  |
          |  |     |     |     |
         4-  +-----+-----+-----+
          |
          +--|-----|-----|-----|--> x_1
             2     3     4     5

For example, the (dimension by number of corners by number of elements) array elemtocoord gives the coordinates of the corners of each element.

julia> topology.elemtocoord
2×4×6 Array{Int64,3}:
[:, :, 1] =
 2  3  2  3
 4  4  5  5

[:, :, 2] =
 3  4  3  4
 4  4  5  5

[:, :, 3] =
 2  3  2  3
 5  5  6  6

[:, :, 4] =
 3  4  3  4
 5  5  6  6

[:, :, 5] =
 4  5  4  5
 5  5  6  6

[:, :, 6] =
 4  5  4  5
 4  4  5  5

Note that the corners are listed in Cartesian order.

The (number of faces by number of elements) array elemtobndy gives the boundary number for each face of each element. A zero will be given for connected faces.

julia> topology.elemtobndy
4×6 Array{Int64,2}:
 1  0  1  0  0  0
 0  0  0  0  2  2
 0  0  0  0  0  0
 0  0  0  0  0  0

Note that the faces are listed in Cartesian order.

Todo

We may/probably want to unify to a single Topology type which has different constructors since all the topologies we currently have are essentially the same

source

VanillaAtmosDiscretizations

CLIMA.CLIMAAtmosDycore.VanillaAtmosDiscretizations.VanillaAtmosDiscretization — Type.

VanillaAtmosDiscretization{nmoist, ntrace}(grid; gravity = true,
viscosity = 0)

Given a 'grid <: AbstractGrid' this construct all the data necessary to run a vanilla discontinuous Galerkin discretization of the the compressible Euler equations with nmoist moisture variables and ntrace tracer variables. If the boolean keyword argument gravity is true then gravity is used otherwise it is not. Isotropic viscosity can be used if viscosity is set to a positive constant.

source

CLIMA.CLIMAAtmosDycore.VanillaAtmosDiscretizations.estimatedt — Function.

estimatedt(disc::VanillaAtmosDiscretization, Q::AtmosStateArray)

Given a discretization disc and a state Q compute an estimate for the time step

Todo

This estimate is currently very conservative, needs to be revisited

source

LSRKmethods

CLIMA.CLIMAAtmosDycore.LSRKmethods.LSRK — Type.

LSRK(f, Q; dt, t0 = 0)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

Q̇ = f(Q)

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order, low-storage, Runge–Kutta scheme of Carpenter and Kennedy (1994) (in their notation (5,4) 2N-Storage RK scheme).

References

@TECHREPORT{CarpenterKennedy1994,
  author = {M.~H. Carpenter and C.~A. Kennedy},
  title = {Fourth-order {2N-storage} {Runge-Kutta} schemes},
  institution = {National Aeronautics and Space Administration},
  year = {1994},
  number = {NASA TM-109112},
  address = {Langley Research Center, Hampton, VA},
}

source