MoistThermodynamics Module

The MoistThermodynamics module provides all thermodynamic functions needed for the atmosphere and functions shared across model components. The functions are general for a moist atmosphere that includes suspended cloud condensate in the working fluid; the special case of a dry atmosphere is obtained for zero specific humidities (or simply by omitting the optional specific humidity arguments in the functions that are needed for a dry atmosphere). The general formulation assumes that there are tracers for the total water specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice to characterize the thermodynamic state and composition of moist air.

There are several types of functions:

Equation of state (ideal gas law):

air_pressure

Specific gas constant and isobaric and isochoric specific heats of moist air:

gas_constant_air

cp_m

cv_m

Specific latent heats of vaporization, fusion, and sublimation:

latent_heat_vapor

latent_heat_fusion

latent_heat_sublim

Saturation vapor pressure and specific humidity over liquid and ice:

sat_vapor_press_liquid

sat_vapor_press_ice

sat_shum

Functions computing energies and inverting them to obtain temperatures

total_energy

internal_energy

air_temperature

Functions to compute temperatures and partitioning of water into phases in thermodynamic equilibrium (when Gibbs' phase rule implies that the entire thermodynamic state of moist air, including the liquid and ice specific humidities, can be calculated from the 3 thermodynamic state variables, such as energy, pressure, and total specific humidity)

liquid_fraction (fraction of condensate that is liquid)

saturation_adjustment (compute temperature from energy, density, and total specific humidity)

Auxiliary functions for diagnostic purposes, e.g., other thermodynamic quantities

liquid_ice_pottemp (liquid-ice potential temperature)

A moist dynamical core that assumes equilibrium thermodynamics can be obtained from a dry dynamical core with total energy as a prognostic variable by including a tracer for the total specific humidity q_tot, using the functions, e.g., for the energies in the module, and computing the temperature T and the liquid and ice specific humidities (q.liq and q.ice) from the internal energy e_int by saturation adjustment:

T = saturation_adjustment(e_int, ρ, q_tot);
q = PhasePartition_equil(T, ρ, q_tot);

here, ρ is the density of the moist air, and the internal energy e_int = e_tot - e_kin - geopotential is the total energy e_tot minus kinetic energy e_kin and potential energy geopotential (all energies per unit mass). No changes to the "right-hand sides" of the dynamical equations are needed for a moist dynamical core that supports clouds, as long as they do not precipitate. Additional source-sink terms arise from precipitation.

Schematically, the workflow in such a core would look as follows:

# initialize
geopotential = grav * z
q_tot          = ...
ρ            = ...

(u, v, w)    = ...
e_kin           = 0.5 * (u^2 + v^2 + w^2)

e_tot        = total_energy(e_kin, geopotential, T, q_tot)

do timestep   # timestepping loop

  # advance dynamical variables by a timestep (temperature typically
  # appears in terms on the rhs, such as radiative transfer)
  advance(u, v, w, ρ, e_tot, q_tot)

  # compute internal energy from dynamic variables
  e_int = e_tot - 0.5 * (u^2 + v^2 + w^2) - geopotential

  # compute temperature, pressure and condensate specific humidities,
  T = saturation_adjustment(e_int, ρ, q_tot);
  q = PhasePartition_equil(T, ρ, q_tot);
  p = air_pressure(T, ρ, q)

end

For a dynamical core that additionally uses the liquid and ice specific humidities q.liq and q.ice as prognostic variables, and thus explicitly allows the presence of non-equilibrium phases such as supercooled water, the saturation adjustment in the above workflow is replaced by a direct calculation of temperature and pressure:

T = air_temperature(e_int, q)
p = air_pressure(T, ρ, q)

Functions

CLIMA.MoistThermodynamics.PhasePartition — Type

PhasePartition

Represents the mass fractions of the moist air mixture.

Constructors

PhasePartition(q_tot::Real[, q_liq::Real[, q_ice::Real]])
PhasePartition(ts::ThermodynamicState)

See also PhasePartition_equil

Fields

tot

total specific humidity

liq

liquid water specific humidity (default: 0)

ice

ice specific humidity (default: 0)

CLIMA.MoistThermodynamics.PhasePartition_equil — Function

PhasePartition_equil(T, ρ, q_tot)

Partition the phases in equilibrium, returning a PhasePartition object using the liquid_fraction function where

T temperature

ρ (moist-)air density

q_tot total specific humidity

The residual q.tot - q.liq - q.ice is the vapor specific humidity.

CLIMA.MoistThermodynamics.ThermodynamicState — Type

ThermodynamicState{FT}

A thermodynamic state, which can be initialized for various thermodynamic formulations (via its sub-types). All ThermodynamicState's have access to functions to compute all other thermodynamic properties.

CLIMA.MoistThermodynamics.PhaseDry — Type

PhaseDry{FT} <: ThermodynamicState

A dry thermodynamic state (q_tot = 0).

Constructors

PhaseDry(e_int, ρ)

Fields

e_int

internal energy

ρ

density of dry air

CLIMA.MoistThermodynamics.PhaseEquil — Type

PhaseEquil{FT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic equilibrium (therefore, saturation adjustment may be needed).

Constructors

PhaseEquil(e_int, ρ, q_tot)

Fields

e_int

internal energy

ρ

density of air (potentially moist)

q_tot

total specific humidity

T

temperature: computed via saturation_adjustment

CLIMA.MoistThermodynamics.PhaseNonEquil — Type

PhaseNonEquil{FT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic non-equilibrium (therefore, temperature can be computed directly).

Constructors

PhaseNonEquil(e_int, q::PhasePartition, ρ)

Fields

e_int

internal energy

ρ

density of air (potentially moist)

q

phase partition

CLIMA.MoistThermodynamics.TemperatureSHumEquil — Function

TemperatureSHumEquil(T, p, q_tot)

Constructs a PhaseEquil thermodynamic state from temperature.

T temperature

p pressure

q_tot total specific humidity

CLIMA.MoistThermodynamics.LiquidIcePotTempSHumEquil — Function

LiquidIcePotTempSHumEquil(θ_liq_ice, ρ, q_tot)

Constructs a PhaseEquil thermodynamic state from:

θ_liq_ice liquid-ice potential temperature

ρ (moist-)air density

q_tot total specific humidity

tol tolerance for saturation adjustment

maxiter maximum iterations for saturation adjustment

CLIMA.MoistThermodynamics.LiquidIcePotTempSHumNonEquil — Function

LiquidIcePotTempSHumNonEquil(θ_liq_ice, ρ, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

θ_liq_ice liquid-ice potential temperature

ρ (moist-)air density

q_pt phase partition

and, optionally

tol tolerance for non-linear equation solve

maxiter maximum iterations for non-linear equation solve

CLIMA.MoistThermodynamics.LiquidIcePotTempSHumNonEquil_given_pressure — Function

LiquidIcePotTempSHumNonEquil_given_pressure(θ_liq_ice, p, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

θ_liq_ice liquid-ice potential temperature

p pressure

q_pt phase partition

CLIMA.MoistThermodynamics.air_density — Function

air_density(T, p[, q::PhasePartition])

The (moist-)air density from the equation of state (ideal gas law) where

T air temperature

p pressure

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_density(ts::ThermodynamicState)

The (moist-)air density, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.air_pressure — Function

air_pressure(T::FT, T∞::FT, p∞::FT, ::DryAdiabaticProcess)

The air pressure for an isentropic process, where

T temperature

T∞ ambient temperature

p∞ ambient pressure

air_pressure(T, ρ[, q::PhasePartition])

The air pressure from the equation of state (ideal gas law) where

T air temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_pressure(ts::ThermodynamicState)

The air pressure from the equation of state (ideal gas law), given a thermodynamic state ts.

CLIMA.MoistThermodynamics.air_temperature — Function

air_temperature(p::FT, θ::FT, Φ::FT, ::DryAdiabaticProcess)

The air temperature for an isentropic process, where

p pressure

θ potential temperature

air_temperature(e_int, q::PhasePartition)

The air temperature, where

e_int internal energy per unit mass

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_temperature(ts::ThermodynamicState)

The air temperature, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.air_temperature_from_liquid_ice_pottemp — Function

air_temperature_from_liquid_ice_pottemp(θ_liq_ice, ρ, q::PhasePartition)

The temperature given

θ_liq_ice liquid-ice potential temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

CLIMA.MoistThermodynamics.cp_m — Function

cp_m([q::PhasePartition])

The isobaric specific heat capacity of moist air where, optionally,

q PhasePartition. Without this argument, the results are for dry air.

cp_m(ts::ThermodynamicState)

The isobaric specific heat capacity of moist air, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.cv_m — Function

cv_m([q::PhasePartition])

The isochoric specific heat capacity of moist air where optionally,

q PhasePartition. Without this argument, the results are for dry air.

cv_m(ts::ThermodynamicState)

The isochoric specific heat capacity of moist air given a thermodynamic state ts.

CLIMA.MoistThermodynamics.dry_pottemp — Function

dry_pottemp(T, ρ[, q::PhasePartition])

The dry potential temperature where

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

dry_pottemp(ts::ThermodynamicState)

The dry potential temperature, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.exner — Function

exner(T, ρ[, q::PhasePartition)])

The Exner function where

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

exner(ts::ThermodynamicState)

The Exner function, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.gas_constant_air — Function

gas_constant_air([q::PhasePartition])

The specific gas constant of moist air given

q PhasePartition. Without this argument, the results are for dry air.

gas_constant_air(ts::ThermodynamicState)

The specific gas constant of moist air given a thermodynamic state ts.

CLIMA.MoistThermodynamics.Ice — Type

Ice <: Phase

An ice phase, to dispatch over saturation_vapor_pressure and q_vap_saturation_generic.

CLIMA.MoistThermodynamics.internal_energy — Function

internal_energy(T[, q::PhasePartition])

The internal energy per unit mass, given a thermodynamic state ts or

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

internal_energy(ts::ThermodynamicState)

The internal energy per unit mass, given a thermodynamic state ts.

internal_energy(ρ::FT, ρe::FT, ρu::AbstractVector{FT}, e_pot::FT)

The internal energy per unit mass, given

ρ (moist-)air density

ρe total energy per unit volume

ρu momentum vector

e_pot potential energy (e.g., gravitational) per unit mass

CLIMA.MoistThermodynamics.internal_energy_sat — Function

internal_energy_sat(T, ρ, q_tot)

The internal energy per unit mass in thermodynamic equilibrium at saturation where

T temperature

ρ (moist-)air density

q_tot total specific humidity

internal_energy_sat(ts::ThermodynamicState)

The internal energy per unit mass in thermodynamic equilibrium at saturation, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.latent_heat_fusion — Function

latent_heat_fusion(T::FT) where {FT<:Real}

The specific latent heat of fusion where

T temperature

latent_heat_fusion(ts::ThermodynamicState)

The specific latent heat of fusion given a thermodynamic state ts.

CLIMA.MoistThermodynamics.latent_heat_sublim — Function

latent_heat_sublim(T::FT) where {FT<:Real}

The specific latent heat of sublimation where

T temperature

latent_heat_sublim(ts::ThermodynamicState)

The specific latent heat of sublimation given a thermodynamic state ts.

CLIMA.MoistThermodynamics.latent_heat_vapor — Function

latent_heat_vapor(T::FT) where {FT<:Real}

The specific latent heat of vaporization where

T temperature

latent_heat_vapor(ts::ThermodynamicState)

The specific latent heat of vaporization given a thermodynamic state ts.

CLIMA.MoistThermodynamics.Liquid — Type

Liquid <: Phase

A liquid phase, to dispatch over saturation_vapor_pressure and q_vap_saturation_generic.

CLIMA.MoistThermodynamics.liquid_fraction — Function

liquid_fraction(T[, q::PhasePartition])

The fraction of condensate that is liquid where

T temperature

q PhasePartition

If q.liq or q.ice are nonzero, the liquid fraction is computed from them.

Otherwise, phase equilibrium is assumed so that the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze.

liquid_fraction(ts::ThermodynamicState)

The fraction of condensate that is liquid given a thermodynamic state ts.

CLIMA.MoistThermodynamics.liquid_ice_pottemp — Function

liquid_ice_pottemp(T, ρ, q::PhasePartition)

The liquid-ice potential temperature where

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

liquid_ice_pottemp(ts::ThermodynamicState)

The liquid-ice potential temperature, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.liquid_ice_pottemp_sat — Function

liquid_ice_pottemp_sat(T, ρ[, q::PhasePartition])

The saturated liquid ice potential temperature where

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

liquid_ice_pottemp_sat(T, ρ, q_tot)

The saturated liquid ice potential temperature where

T temperature

ρ (moist-)air density

q_tot total specific humidity

liquid_ice_pottemp_sat(ts::ThermodynamicState)

The liquid potential temperature given a thermodynamic state ts.

CLIMA.MoistThermodynamics.gas_constants — Function

(R_m, cp_m, cv_m, γ_m) = gas_constants([q::PhasePartition])

Wrapper to compute all gas constants at once, where optionally,

q PhasePartition. Without this argument, the results are for dry air.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

(R_m, cp_m, cv_m, γ_m) = gas_constants(ts::ThermodynamicState)

Wrapper to compute all gas constants at once, given a thermodynamic state ts.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

CLIMA.MoistThermodynamics.saturation_adjustment — Function

saturation_adjustment(e_int, ρ, q_tot)

Compute the temperature that is consistent with

e_int internal energy

ρ (moist-)air density

q_tot total specific humidity

tol tolerance for non-linear equation solve

maxiter maximum iterations for non-linear equation solve

by finding the root of

$e_int - internal_energy_sat(T,ρ,q_tot) = 0$

using Newtons method with analytic gradients.

See also saturation_adjustment.

CLIMA.MoistThermodynamics.saturation_excess — Function

saturation_excess(T, ρ, q::PhasePartition)

The saturation excess in equilibrium where

T temperature

ρ (moist-)air density

q PhasePartition

The saturation excess is the difference between the total specific humidity q.tot and the saturation specific humidity in equilibrium, and it is defined to be nonzero only if this difference is positive.

saturation_excess(ts::ThermodynamicState)

Compute the saturation excess in equilibrium, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.q_vap_saturation — Function

q_vap_saturation(T, ρ[, q::PhasePartition])

Compute the saturation specific humidity, given

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition

If the PhasePartition q is given, the saturation specific humidity is that of a mixture of liquid and ice, computed in a thermodynamically consistent way from the weighted sum of the latent heats of the respective phase transitions (Pressel et al., JAMES, 2015). That is, the saturation vapor pressure and from it the saturation specific humidity are computed from a weighted mean of the latent heats of vaporization and sublimation, with the weights given by the fractions of condensates q.liq/(q.liq + q.ice) and q.ice/(q.liq + q.ice) that are liquid and ice, respectively.

If the PhasePartition q is not given, or has zero liquid and ice specific humidities, the saturation specific humidity is that over a mixture of liquid and ice, with the fraction of liquid given by temperature dependent liquid_fraction(T) and the fraction of ice by the complement 1 - liquid_fraction(T).

q_vap_saturation(ts::ThermodynamicState)

Compute the saturation specific humidity, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.q_vap_saturation_generic — Function

q_vap_saturation_generic(T, ρ[; phase=Liquid()])

Compute the saturation specific humidity over a plane surface of condensate, given

T temperature

ρ (moist-)air density

and, optionally,

Liquid() indicating condensate is liquid

Ice() indicating condensate is ice

CLIMA.MoistThermodynamics.saturation_vapor_pressure — Function

saturation_vapor_pressure(T, Liquid())

Return the saturation vapor pressure over a plane liquid surface at temperature T.

saturation_vapor_pressure(T, Ice())

Return the saturation vapor pressure over a plane ice surface at temperature T.

saturation_vapor_pressure(T, LH_0, Δcp)

Compute the saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The Clausius-Clapeyron relation

dlog(p_v_sat)/dT = [LH_0 + Δcp * (T-T_0)]/(R_v*T^2)

is integrated from the triple point temperature T_triple, using Kirchhoff's relation

L = LH_0 + Δcp * (T - T_0)

for the specific latent heat L with constant isobaric specific heats of the phases. The linear dependence of the specific latent heat on temperature T allows analytic integration of the Clausius-Clapeyron relation to obtain the saturation vapor pressure p_v_sat as a function of the triple point pressure press_triple.

CLIMA.MoistThermodynamics.soundspeed_air — Function

soundspeed_air(T[, q::PhasePartition])

The speed of sound in unstratified air, where

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

soundspeed_air(ts::ThermodynamicState)

The speed of sound in unstratified air given a thermodynamic state ts.

CLIMA.MoistThermodynamics.specific_volume — Function

specific_volume(ts::ThermodynamicState)

The (moist-)air specific volume, given a thermodynamic state ts.

CLIMA.MoistThermodynamics.total_energy — Function

total_energy(e_kin, e_pot, T[, q::PhasePartition])

The total energy per unit mass, given

e_kin kinetic energy per unit mass

e_pot potential energy per unit mass

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

total_energy(e_kin, e_pot, ts::ThermodynamicState)

The total energy per unit mass given a thermodynamic state ts.

CLIMA.MoistThermodynamics.virtual_pottemp — Function

virtual_pottemp(T, ρ[, q::PhasePartition])

The virtual temperature where

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

virtual_pottemp(ts::ThermodynamicState)

The virtual potential temperature, given a thermodynamic state ts.