DGmethods_old

SpaceMethods

Set of abstract types for spatial discretizations

Types

CLIMA.SpaceMethods.AbstractSpaceMethod — Type

AbstractSpaceMethod

Supertype for spatial discretizations

CLIMA.SpaceMethods.AbstractDGMethod — Type

AbstractDGMethod <: AbstractSpaceMethod

Supertype for discontinuous Galerkin spatial discretizations

DGBalanceLawDiscretizations

CLIMA.DGBalanceLawDiscretizations — Module

DG Balance Law Discretizations module. Attempts to provide a reasonable implementation of a discontinuous Galerkin method (in weak form) on tensor product quadrilateral (2D) and hexahedral (3D) elements for balance laws of the form

\[q_{,t} + Σ_{i=1,...d} F_{i,i} = s\]

where $q$ is the state vector, $F$ is the flux function, and $s$ is the source function. $F$ includes both the "inviscid" and "viscous" fluxes. Note that this is a space only discretization, time must be advanced using some ordinary differential equations methods; see ODESolvers.

The flux function F_{i} is taken to be of the form:

\[F_{i} := F_{i}(q, σ; a)\\ σ = H(q, ∇G(q; a); a)\]

where $a$ is a set of parameters and viscous terms enter through $σ$

The source term is of the form:

\[s := s(q; a)\]

In the code and docs the following terminology is used:

$q$ is referred to as the state

$σ$ is the viscous state

$a$ is the auxiliary state

$F$ is the physical flux

$H$ is the viscous transform

$G$ is the gradient transform

Much of the notation used in this module follows Hesthaven and Warburton (2008).

Note

Currently all the functions take the same parameters and the gradient transform can take a user-specified subset of the state vector.

References

@BOOK{HesthavenWarburton2008,
  title = {Nodal Discontinuous {G}alerkin Methods: {A}lgorithms, Analysis,
           and Applications},
  publisher = {Springer-Verlag New York},
  year = {2008},
  author = {Hesthaven, Jan S. and Warburton, Tim},
  volume = {54},
  series = {Texts in Applied Mathematics},
  doi = {10.1007/978-0-387-72067-8}
}

Types/Constructors

CLIMA.DGBalanceLawDiscretizations.DGBalanceLaw — Type

DGBalanceLaw <: AbstractDGMethod

This contains the necessary information for a discontinuous Galerkin method for balance laws.

See also: Outer constructor DGBalanceLaw

Fields

grid

computational grid / mesh

nstate

number of state

flux!

physical flux function

numerical_flux!

numerical flux function

numerical_boundary_flux!

numerical boundary flux function

Qvisc

storage for the viscous state

number_gradient_states

number of out states for gradient_transform!

number_viscous_states

number of out states for the viscous_transform!

gradient_transform!

transform from state to variables to take gradient of

viscous_transform!

transform from Q and gradient state to viscous states

viscous_penalty!

penalty for the viscous state computation

viscous_boundary_penalty!

boundary penalty for the viscous state computation (e.g., Dirichlet)

auxstate

auxiliary state array

source!

source function

preodefun!

callback function for before the odefun!

Functions

CLIMA.DGBalanceLawDiscretizations.grad_auxiliary_state! — Function

grad_auxiliary_state!(disc, i, (ix1, ix2, ix3)

Computes the gradient of a the field i of the constant auxiliary state of disc and stores the x1, x2, x3 compoment in fields ix1, ix2, ix3 of constant auxiliary state.

Note

This only computes the element gradient not a DG gradient. If your constant auxiliary state is discontinuous this may or may not be what you want!

CLIMA.DGBalanceLawDiscretizations.indefinite_stack_integral! — Function

indefinite_stack_integral!(disc, f, Q, out_states, [P=disc.auxstate])

Computes an indefinite line integral along the trailing dimension (ξ3 in 3-D and ξ2 in 2-D) up an element stack using state Q

\[∫_{ζ_{0}}^{ζ} f(q; aux, t)\]

and stores the result of the integral in field of P indicated by out_states

The syntax of the integral kernel is:

f(F, Q, aux)

where F is an MVector of length length(out_states), Q and aux are the MVectors for the state and auxiliary state at a single degree of freedom. The function is responsible for filling F.

Requires the isstacked(disc.grid.topology) == true

CLIMA.DGBalanceLawDiscretizations.reverse_indefinite_stack_integral! — Function

reverse_indefinite_stack_integral!(disc, oustate, instate,
                                   [P=disc.auxstate])

reverse previously computed indefinite integral(s) computed with indefinite_stack_integral! to be

\[∫_{ζ}^{ζ_{max}} f(q; aux, t)\]

The states instate[i] is reverse and stored in instate[i].

Requires the isstacked(disc.grid.topology) == true

CLIMA.DGBalanceLawDiscretizations.dof_iteration! — Function

dof_iteration!(dof_fun!::Function, R::MPIStateArray, disc::DGBalanceLaw,
               Q::MPIStateArray)

Iterate over each dof to fill R using the dof_fun!. The syntax of the dof_fun! is

dof_fun!(l_R, l_Q, l_Qvisc, l_aux)

where l_R, l_Q, l_Qvisc, and l_aux are of type MArray filled initially with the values at a single degree of freedom. After the call the values in l_R will be written back to the degree of freedom of R.

Pirated Type Functions

CLIMA.MPIStateArrays.MPIStateArray — Method

MPIStateArray(disc::DGBalanceLaw; nstate=disc.nstate, commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional 'nstate' arguments allows the user to specify a specific number of states. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

CLIMA.MPIStateArrays.MPIStateArray — Method

MPIStateArray(disc::DGBalanceLaw, initialization!::Function; commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

After allocation the MPIStateArray is initialized using the function initialization! which will be called as:

initialization!(Q, x1, x2, x3, aux)

where Q is an MArray with the solution state at a single degree of freedom (DOF) to initialize and (x1, x2, x3) is the coordinate point for the allocation. The auxiliary data the values at the DOF are passed through as an MArray through the aux argument; if disc does not have auxiliary data then the length of the MArray will be zero.

Note

Q is undef at start the function (i.e., not initialized to zero)

Note

Modifications of the aux array will be discarded.

Todo

Remove host and device data transfers.

CLIMA.SpaceMethods.odefun! — Function

odefun!(disc::AbstractSpaceMethod, dQ, Q, t; increment)

Evaluates the right-hand side of the spatial discretization defined by disc at time t with state Q. The result is either added into dQ if increment is true or stored in dQ if it is false. Namely, the semi-discretization is of the form $\dot{Q} = F(Q, t)$ and after the call dQ += F(Q, t) if increment == true or dQ = F(Q, t) if increment == false

Note

There is no generic implementation of this function. This must be implemented for each subtype of AbstractSpaceMethod

Kernels

CLIMA.DGBalanceLawDiscretizations.volumerhs! — Function

volumerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{nviscstate},
           ::Val{nauxstate}, flux!, source!, rhs, Q, Qvisc, auxstate,
           vgeo, t, D, elems) where {dim, N, nstate, nviscstate,

Computational kernel: Evaluate the volume integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

CLIMA.DGBalanceLawDiscretizations.facerhs! — Function

facerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{nviscstate},
         ::Val{nauxstate}, numerical_flux!,
         numerical_boundary_flux!, rhs, Q, Qvisc, auxstate,
         vgeo, sgeo, t, vmap⁻, vmap⁺, elemtobndy,
         elems) where {dim, N, nstate, nviscstate, nauxstate}

Computational kernel: Evaluate the surface integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

CLIMA.DGBalanceLawDiscretizations.initauxstate! — Function

initauxstate!(::Val{dim}, ::Val{N}, ::Val{nauxstate}, auxstatefun!,
              auxstate, vgeo, elems) where {dim, N, nauxstate}

Computational kernel: Initialize the auxiliary state

See DGBalanceLaw for usage.

CLIMA.DGBalanceLawDiscretizations.elem_grad_field! — Function

elem_grad_field!(::Val{dim}, ::Val{N}, ::Val{nstate}, Q, vgeo, D, elems, s,
                 sx, sy, sz) where {dim, N, nstate}

Computational kernel: Compute the element gradient of state s of Q and store it in sx, sy, and sz of Q.

Warning

This does not compute a DG gradient, but only over the element. If $Q_s$ is discontinuous you may want to consider another approach.

CLIMA.DGBalanceLawDiscretizations.knl_dof_iteration! — Function

knl_dof_iteration!(::Val{dim}, ::Val{N}, ::Val{nRstate}, ::Val{nstate},
                   ::Val{nviscstate}, ::Val{nauxstate}, dof_fun!, R, Q,
                   QV, auxstate, elems) where {dim, N, nRstate, nstate,
                                               nviscstate, nauxstate}

Computational kernel: fill postprocessing array

See DGBalanceLaw for usage.

CLIMA.DGBalanceLawDiscretizations.knl_indefinite_stack_integral! — Function

knl_indefinite_stack_integral!(::Val{dim}, ::Val{N}, ::Val{nstate},
                                        ::Val{nauxstate}, ::Val{nvertelem},
                                        int_knl!, Q, auxstate, vgeo, Imat,
                                        elems, ::Val{outstate}
                                       ) where {dim, N, nstate, nauxstate,
                                                outstate, nvertelem}

Computational kernel: compute indefinite integral along the vertical stack

See DGBalanceLaw for usage.

Numerical Fluxes

CLIMA.DGBalanceLawDiscretizations.NumericalFluxes.rusanov! — Function

rusanov!(F::MArray, nM, QM, QVM, auxM, QP, QVP, auxP, t, flux!, wavespeed,
         [computeQjump!])

Calculate the Rusanov (aka local Lax-Friedrichs) numerical flux given the plus and minus side states/viscous states QP/QVP and QM/QVM using the physical flux function flux! and wavespeed calculation.

The flux! has almost the same calling convention as flux! from DGBalanceLaw.

The function wavespeed should return the maximum wavespeed for a state and is called as wavespeed(nM, QM, auxM, t) and wavespeed(nM, QP, auxP, t) where nM is the outward unit normal for the minus side.

When present computeQjump!(ΔQ, QM, auxM, QP, auxP) will be called after so that the user specify the value to use for QM - QP; this is useful for correcting Q to include discontinuous reference states.

CLIMA.DGBalanceLawDiscretizations.NumericalFluxes.rusanov_boundary_flux! — Function

rusanov_boundary_flux!(F::MArray{Tuple{nstate}}, nM, QM, QVM, auxM, QP, QVP,
                       auxP, bctype, t, flux!, bcstate!, wavespeed,
                       computeQjump! = nothing) where {nstate}

The function bcstate! is used to calculate the plus side state for the boundary condition bctype. The calling convention is:

bcstate!(QP, QVP, auxP, nM, QM, QVM, auxM, bctype, t)

where QP, QVP, and auxP are the plus side state, viscous state, and auxiliary state to be filled from the given data; other arguments should not be modified.