InputOutput

InputOutput

VTK

Functions

CLIMA.VTK.writevtk — Function.

writevtk(prefix, Q::MPIStateArray, disc::DGBalanceLaw [, fieldnames])

Write a vtk file for all the fields in the state array Q using geometry and connectivity information from disc.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames; if not specified the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

source

writevtk(prefix, Q::MPIStateArray, disc::DGBalanceLaw, fieldnames,
         auxstate::MPIStateArray, auxfieldnames)

Write a vtk file for all the fields in the state array Q and auxiliary state auxstate using geometry and connectivity information from disc.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames and auxfieldnames.

If fieldnames === nothing then the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

If auxfieldnames === nothing then the fields names will be "aux1" through "auxk" where k is the number of states in auxstate, i.e., k = size(auxstate,2).

source

CLIMA.VTK.writevtk_helper — Function.

writevtk_helper(prefix, vgeo::Array, Q::Array, grid, fieldnames)

Internal helper function for writevtk

source