TurbulenceConvection

TurbulenceConvection

Grids

CLIMA.TurbulenceConvection.Grids.Grid — Type.

Grid{T}

A simple 1-dimensional uniform grid of type T for finite difference method.

source

CLIMA.TurbulenceConvection.Grids.over_elems — Function.

over_elems(grid::Grid)

Get the range of indexes to traverse real and ghost grid elements

source

CLIMA.TurbulenceConvection.Grids.over_elems_real — Function.

over_elems_real(grid::Grid)

Get the range of indexes to traverse only real grid elements

source

CLIMA.TurbulenceConvection.Grids.over_elems_ghost — Function.

over_elems_ghost(grid::Grid)

Get the range of indexes to traverse only ghost grid elements

source

CLIMA.TurbulenceConvection.Grids.first_elem_above_surface — Function.

first_elem_above_surface(grid::Grid)

Get the first element index above the surface

source

CLIMA.TurbulenceConvection.Grids.get_z — Function.

get_z(grid::Grid, k::Int)

Get the z-coordinate given element index

source

Grid Operators

CLIMA.TurbulenceConvection.GridOperators.∇_z — Function.

∇_z(f, grid::Grid)

Computes the local derivative of field f

source

CLIMA.TurbulenceConvection.GridOperators.Δ_z — Function.

Δ_z(f, grid::Grid)

Computes the local Laplacian of field f

source

CLIMA.TurbulenceConvection.GridOperators.adv_upwind — Function.

adv_upwind(ϕ, u, grid::Grid)

Local upwind advection operator u⋅∇ϕ. This operator is stable but numerically diffusive.

source

State Vector

CLIMA.TurbulenceConvection.StateVecs.StateVec — Type.

StateVec{T, I, NT}

A state vector containing the number of subdomains, n_subdomains, a NamedTuple variable mapper, a tuple of the variable names, and a vector of vectors, containing the values for all of the variables.

source

CLIMA.TurbulenceConvection.StateVecs.over_sub_domains — Function.

over_sub_domains(state_vec::StateVec)

Get list of indexes from 1 to the maximum subdomain size.

source

over_sub_domains(state_vec::StateVec, j::Int)

Get list of indexes from 1 to the maximum subdomain size, except the given index.

source

over_sub_domains(state_vec::StateVec, name::Symbol)

Get list of indexes over all subdomains for variable name.

source

CLIMA.TurbulenceConvection.StateVecs.Slice — Type.

Slice{I} <: AbstractSlice{I}

A slice struct used to slice the state vector along the grid-element dimension. This is used to as an API to pass slices into local derivative/interpolation routines.

source

Boundary Conditions

CLIMA.TurbulenceConvection.BoundaryConditions.Dirichlet! — Function.

Dirichlet!(sv::StateVec, name::Symbol, val, grid, ::Bottom, i_sd=1)

Apply Dirichlet boundary conditions at the bottom of the domain

source

Dirichlet!(sv::StateVec, name::Symbol, val, grid, ::Top, i_sd=1)

Apply Dirichlet boundary conditions at the bottom of the domain

source

CLIMA.TurbulenceConvection.BoundaryConditions.Neumann! — Function.

Neumann!(sv::StateVec, name::Symbol, val, grid, ::Bottom, i_sd=1)

Apply Neumann boundary conditions at the bottom of the domain

source

Neumann!(sv::StateVec, name::Symbol, val, grid, ::Top, i_sd=1)

Apply Neumann boundary conditions at the bottom of the domain

source

CLIMA.TurbulenceConvection.BoundaryConditions.Top — Type.

Top<:BoundaryLocation

Location to apply boundary condition

source

CLIMA.TurbulenceConvection.BoundaryConditions.Bottom — Type.

Bottom<:BoundaryLocation

Location to apply boundary condition

source