Utilities

Utilities

Root solvers

CLIMA.Utilities.RootSolvers — Module.

RootSolvers

Module containing functions for solving roots of non-linear equations. The returned result is a tuple of the root and a Bool indicating convergence.

find_zero(f::F,
           x_0::T,
           x_1::T,
           args::Tuple,
           iter_params::IterParams{R, Int},
           method::RootSolvingMethod
           )::Tuple{T, Bool} where {F, R, T <: Union{R, AbstractArray{R}}}

Interface

find_zero compute x^* such that f(x^*) = 0

Arguments

f equation roots function, where f is callable via f(x, args...)

x_0, x_1 initial guesses

IterParams struct containing absolute tolerance on f(x^*) and maximum iterations

RootSolvingMethod Algorithm to solve roots of the equation:

SecantMethod Secant method

RegulaFalsiMethod Regula Falsi Method

Single example

julia> using RootSolvers
x_0 = 0.0
x_1 = 1.0
f(x, y) = x^2 - y
x_star2 = 10000.0
args = Tuple(x_star2)
x_star = sqrt(x_star2)
tol_abs = 1.0e-3
iter_max = 100

x_root, converged = find_zero(f,
                              x_0,
                              x_1,
                              args,
                              IterParams(tol_abs, iter_max),
                              SecantMethod())

Broadcast example

To broadcast, wrap arguments that need not be broadcasted in a Ref.

julia> using RootSolvers
x_0 = rand(5,5)
x_1 = rand(5,5).+2.0
f(x, y) = x^2 - y
x_star2 = 10000.0
args = Tuple(x_star2)
x_star = sqrt(x_star2)
tol_abs = 1.0e-3
iter_max = 100
x_root, converged = find_zero.(f,
                               x_0,
                               x_1,
                               Ref(args),
                               Ref(IterParams(tol_abs, iter_max)),
                               Ref(SecantMethod()))

source

CLIMA.Utilities.RootSolvers.find_zero — Function.

find_zero(f, x_0, x_1, args, iter_params, SecantMethod)

Solves the root equation, f, using Secant method. See RootSolvers for more information.

source

find_zero(f, x_0, x_1, args, iter_params, RegulaFalsiMethod)

Solves the root equation, f, using Regula Falsi method. See RootSolvers for more information.

source

Moist thermodynamics

CLIMA.Utilities.MoistThermodynamics — Module.

MoistThermodynamics

Moist thermodynamic functions, e.g., for air pressure (atmosphere equation of state), latent heats of phase transitions, saturation vapor pressures, and saturation specific humidities

source

CLIMA.Utilities.MoistThermodynamics.gas_constant_air — Function.

gas_constant_air([q_t=0, q_l=0, q_i=0])

Return the specific gas constant of moist air, given the total specific humidity q_t, and, optionally, the liquid specific humidity q_l, and the ice specific humidity q_i. When no input argument is given, it returns the specific gas constant of dry air.

source

CLIMA.Utilities.MoistThermodynamics.air_pressure — Function.

air_pressure(T, ρ[, q_t=0, q_l=0, q_i=0])

Return the air pressure from the equation of state (ideal gas law), given the air temperature T, the (moist-)air density ρ, and, optionally, the total specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i. Without the specific humidity arguments, it returns the air pressure from the equation of state of dry air.

source

CLIMA.Utilities.MoistThermodynamics.air_density — Function.

air_density(T, p[, q_t=0, q_l=0, q_i=0])

Return the (moist-)air density from the equation of state (ideal gas law), given the air temperature T, the pressure p, and, optionally, the total specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i. Without the specific humidity arguments, it returns the air density from the equation of state of dry air.

source

CLIMA.Utilities.MoistThermodynamics.cp_m — Function.

cp_m([q_t=0, q_l=0, q_i=0])

Return the isobaric specific heat capacity of moist air, given the total water specific humidity q_t, liquid specific humidity q_l, and ice specific humidity q_i. Without the specific humidity arguments, it returns the isobaric specific heat capacity of dry air.

source

CLIMA.Utilities.MoistThermodynamics.cv_m — Function.

cv_m([q_t=0, q_l=0, q_i=0])

Return the isochoric specific heat capacity of moist air, given the total water specific humidity q_t, liquid specific humidity q_l, and ice specific humidity q_i. Without the specific humidity arguments, it returns the isochoric specific heat capacity of dry air.

source

CLIMA.Utilities.MoistThermodynamics.air_temperature — Function.

air_temperature(e_int[, q_t=0, q_l=0, q_i=0])

Return the air temperature, given the internal energy e_int per unit mass, and, optionally, the total specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i.

source

CLIMA.Utilities.MoistThermodynamics.internal_energy — Function.

internal_energy(T[, q_t=0, q_l=0, q_i=0])

Return the internal energy per unit mass, given the temperature T, and, optionally, the total specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i.

source

CLIMA.Utilities.MoistThermodynamics.internal_energy_sat — Function.

internal_energy_sat(T, ρ, q_t)

Return the internal energy per unit mass in thermodynamic equilibrium at saturation, given the temperature T, (moist-)air density ρ, and total specific humidity q_t.

source

CLIMA.Utilities.MoistThermodynamics.total_energy — Function.

total_energy(KE, PE, T[, q_t=0, q_l=0, q_i=0])

Return the total energy per unit mass, given the kinetic energy per unit mass KE, the potential energy per unit mass PE, the temperature T, and, optionally, the total specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i.

source

CLIMA.Utilities.MoistThermodynamics.latent_heat_vapor — Function.

latent_heat_vapor(T)

Return the specific latent heat of vaporization at temperature T.

source

CLIMA.Utilities.MoistThermodynamics.latent_heat_sublim — Function.

latent_heat_sublim(T)

Return the specific latent heat of sublimation at temperature T.

source

CLIMA.Utilities.MoistThermodynamics.latent_heat_fusion — Function.

latent_heat_fusion(T)

Return the specific latent heat of fusion at temperature T.

source

CLIMA.Utilities.MoistThermodynamics.latent_heat_generic — Function.

latent_heat_generic(T, LH_0, cp_diff)

Return the specific latent heat of a generic phase transition between two phases using Kirchhoff's relation.

The latent heat computation assumes constant isobaric specifc heat capacities of the two phases. T is the temperature, LH_0 is the latent heat of the phase transition at T_0, and cp_diff is the difference between the isobaric specific heat capacities (heat capacity in the higher-temperature phase minus that in the lower-temperature phase).

source

CLIMA.Utilities.MoistThermodynamics.saturation_vapor_pressure — Function.

`saturation_vapor_pressure(T, Liquid())`

Return the saturation vapor pressure over a plane liquid surface at temperature T.

`saturation_vapor_pressure(T, Ice())`

Return the saturation vapor pressure over a plane ice surface at temperature T.

`saturation_vapor_pressure(T, LH_0, cp_diff)`

Compute the saturation vapor pressure over a plane surface by integration of the Clausius-Clepeyron relation.

The Clausius-Clapeyron relation

dlog(p_vs)/dT = [LH_0 + cp_diff * (T-T_0)]/(R_v*T^2)

is integrated from the triple point temperature T_triple, using Kirchhoff's relation

L = LH_0 + cp_diff * (T - T_0)

for the specific latent heat L with constant isobaric specific heats of the phases. The linear dependence of the specific latent heat on temperature T allows analytic integration of the Clausius-Clapeyron relation to obtain the saturation vapor pressure p_vs as a function of the triple point pressure press_triple.

source

CLIMA.Utilities.MoistThermodynamics.saturation_shum_generic — Function.

saturation_shum_generic(T, ρ[; phase=Liquid()])

Compute the saturation specific humidity over a plane surface of condensate, given the temperature T and the (moist-)air density ρ.

The optional argument phase can be $Liquid()$ or $"ice"$ and indicates the condensed phase.

source

CLIMA.Utilities.MoistThermodynamics.saturation_shum — Function.

saturation_shum(T, ρ[, q_l=0, q_i=0])

Compute the saturation specific humidity, given the temperature T and (moist-)air density ρ.

If the optional liquid, and ice specific humdities q_t and q_l are given, the saturation specific humidity is that over a mixture of liquid and ice, computed in a thermodynamically consistent way from the weighted sum of the latent heats of the respective phase transitions (Pressel et al., JAMES, 2015). That is, the saturation vapor pressure and from it the saturation specific humidity are computed from a weighted mean of the latent heats of vaporization and sublimation, with the weights given by the fractions of condensate q_l/(q_l + q_i) and q_i/(q_l + q_i) that are liquid and ice, respectively.

If the condensate specific humidities q_l and q_i are not given or are both zero, the saturation specific humidity is that over a mixture of liquid and ice, with the fraction of liquid given by temperature dependent liquid_fraction(T) and the fraction of ice by the complement 1 - liquid_fraction(T).

source

CLIMA.Utilities.MoistThermodynamics.saturation_shum_from_pressure — Function.

saturation_shum_from_pressure(ρ, T, p_vs)

Compute the saturation specific humidity, given the ambient air density ρ, temperature T, and the saturation vapor pressure p_vs.

source

CLIMA.Utilities.MoistThermodynamics.liquid_fraction — Function.

liquid_fraction(T[, q_l=0, q_i=0])

Return the fraction of condensate that is liquid.

If the optional input arguments q_l and q_i are not given or are zero, the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze. If q_l or q_i are nonzero, the liquid fraction is computed from them.

source

CLIMA.Utilities.MoistThermodynamics.heaviside — Function.

heaviside(t)

Return the Heaviside step function at t.

source

CLIMA.Utilities.MoistThermodynamics.phase_partitioning_eq! — Function.

phase_partitioning_eq!(q_l, q_i, T, ρ, q_t)

Return the partitioning of the phases in equilibrium.

Given the temperature T and (moist-)air density ρ, phase_partitioning_eq! partitions the total specific humidity q_t into the liquid specific humidity q_l and ice specific humiditiy q_l using the liquid_fraction function. The residual q_t - q_l - q_i is the vapor specific humidity.

source

CLIMA.Utilities.MoistThermodynamics.saturation_adjustment — Function.

saturation_adjustment(e_int, ρ, q_t[, T_init = T_triple])

Return the temperature that is consistent with the internal energy e_int, (moist-)air density ρ, and total specific humidity q_t.

The optional input value of the temperature T_init is taken as the initial value of the saturation adjustment iterations.

source

CLIMA.Utilities.MoistThermodynamics.liquid_ice_pottemp — Function.

liquid_ice_pottemp(T, p[, q_t=0, q_l=0, q_i=0])

Return the liquid-ice potential temperature, given the temperature T, pressure p, total specific humidity q_t, liquid specific humidity q_l, and ice specific humidity q_i.

source