MoistThermodynamics

MoistThermodynamics Module

The MoistThermodynamics module provides all thermodynamic functions needed for the atmosphere and functions shared across model components. The functions are general for a moist atmosphere that includes suspended cloud condensate in the working fluid; the special case of a dry atmosphere is obtained for zero specific humidities (or simply by omitting the optional specific humidity arguments in the functions that are needed for a dry atmosphere). The general formulation assumes that there are tracers for the total water specific humidity q_t, the liquid specific humidity q_l, and the ice specific humidity q_i to characterize the thermodynamic state and composition of moist air.

There are several types of functions:

Equation of state (ideal gas law):

air_pressure

Specific gas constant and isobaric and isochoric specific heats of moist air:

gas_constant_air

cp_m

cv_m

Specific latent heats of vaporization, fusion, and sublimation:

latent_heat_vapor

latent_heat_fusion

latent_heat_sublim

Saturation vapor pressure and specific humidity over liquid and ice:

sat_vapor_press_liquid

sat_vapor_press_ice

sat_shum

Functions computing energies and inverting them to obtain temperatures

total_energy

internal_energy

air_temperature

Functions to compute temperatures and partitioning of water into phases in thermodynamic equilibrium (when Gibbs' phase rule implies that the entire thermodynamic state of moist air, including the liquid and ice specific humidities, can be calculated from the 3 thermodynamic state variables, such as energy, pressure, and total specific humidity)

liquid_fraction (fraction of condensate that is liquid)

saturation_adjustment (compute temperature from energy, density, and total specific humidity)

Auxiliary functions for diagnostic purposes, e.g., other thermodynamic quantities

liquid_ice_pottemp (liquid-ice potential temperature)

A moist dynamical core that assumes equilibrium thermodynamics can be obtained from a dry dynamical core with total energy as a prognostic variable by including a tracer for the total specific humidity q_t, using the functions, e.g., for the energies in the module, and computing the temperature T and the liquid and ice specific humidities q_l and q_i from the internal energy E_int by saturation adjustment:

    T = saturation_adjustment(E_int, ρ, q_t, T_init);
    phase_partitioning_eq!(q_l, q_i, T, ρ, q_t);

here, ρ is the density of the moist air, T_init is an initial temperature guess for the saturation adjustment iterations, and the internal energy E_int = E_tot - KE - geopotential is the total energy E_tot minus kinetic energy KE and potential energy geopotential (all energies per unit mass). No changes to the "right-hand sides" of the dynamical equations are needed for a moist dynamical core that supports clouds, as long as they do not precipitate. Additional source-sink terms arise from precipitation.

Schematically, the workflow in such a core would look as follows:

    # initialize
    geopotential = grav * z
    T_prev       = ...
    q_t          = ...
    ρ            = ...

    (u, v, w)    = ...
    KE           = 0.5 * (u.^2 .+ v.^2 .+ w.^2)

    E_tot        = total_energy(KE, geopotential, T, q_t)

    do timestep   # timestepping loop

      # advance dynamical variables by a timestep (temperature typically
      # appears in terms on the rhs, such as radiative transfer)
      advance(u, v, w, ρ, E_tot, q_t)  

      # compute internal energy from dynamic variables
      E_int = E_tot - 0.5 * (u.^2 .+ v.^2 .+ w.^2) - geopotential

      # compute temperature, pressure and condensate specific humidities,
      # using T_prev as initial condition for iterations
      T = saturation_adjustment(E_int, ρ, q_t, T_prev);
      phase_partitioning_eq!(q_l, q_i, T, ρ, q_t);
      p = air_pressure(T, ρ, q_t, q_l, q_i)

      # update temperature for next timestep
      T_prev = T;  
    end

For a dynamical core that additionally uses the liquid and ice specific humidities q_l and q_i as prognostic variables, and thus explicitly allows the presence of non-equilibrium phases such as supercooled water, the saturation adjustment in the above workflow is replaced by a direct calculation of temperature and pressure:

    T = air_temperature(E_int, q_t, q_l, q_i)
    p = air_pressure(T, ρ, q_t, q_l, q_i)

Functions

CLIMA.MoistThermodynamics.gas_constant_air — Function.

gas_constant_air([q_tot=0, q_liq=0, q_ice=0])

Return the specific gas constant of moist air, given the total specific humidity q_tot, and, optionally, the liquid specific humidity q_liq, and the ice specific humidity q_ice. When no input argument is given, it returns the specific gas constant of dry air.

source

CLIMA.MoistThermodynamics.air_pressure — Function.

air_pressure(T, ρ[, q_tot=0, q_liq=0, q_ice=0])

Return the air pressure from the equation of state (ideal gas law), given the air temperature T, the (moist-)air density ρ, and, optionally, the total specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice. Without the specific humidity arguments, it returns the air pressure from the equation of state of dry air.

source

CLIMA.MoistThermodynamics.air_density — Function.

air_density(T, p[, q_tot=0, q_liq=0, q_ice=0])

Return the (moist-)air density from the equation of state (ideal gas law), given the air temperature T, the pressure p, and, optionally, the total specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice. Without the specific humidity arguments, it returns the air density from the equation of state of dry air.

source

CLIMA.MoistThermodynamics.cp_m — Function.

cp_m([q_tot=0, q_liq=0, q_ice=0])

Return the isobaric specific heat capacity of moist air, given the total water specific humidity q_tot, liquid specific humidity q_liq, and ice specific humidity q_ice. Without the specific humidity arguments, it returns the isobaric specific heat capacity of dry air.

source

CLIMA.MoistThermodynamics.cv_m — Function.

cv_m([q_tot=0, q_liq=0, q_ice=0])

Return the isochoric specific heat capacity of moist air, given the total water specific humidity q_tot, liquid specific humidity q_liq, and ice specific humidity q_ice. Without the specific humidity arguments, it returns the isochoric specific heat capacity of dry air.

source

CLIMA.MoistThermodynamics.air_temperature — Function.

air_temperature(e_int[, q_tot=0, q_liq=0, q_ice=0])

Return the air temperature, given the internal energy e_int per unit mass, and, optionally, the total specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice.

source

CLIMA.MoistThermodynamics.internal_energy — Function.

internal_energy(T[, q_tot=0, q_liq=0, q_ice=0])

Return the internal energy per unit mass, given the temperature T, and, optionally, the total specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice.

source

CLIMA.MoistThermodynamics.internal_energy_sat — Function.

internal_energy_sat(T, ρ, q_tot)

Return the internal energy per unit mass in thermodynamic equilibrium at saturation, given the temperature T, (moist-)air density ρ, and total specific humidity q_tot.

source

CLIMA.MoistThermodynamics.total_energy — Function.

total_energy(KE, PE, T[, q_tot=0, q_liq=0, q_ice=0])

Return the total energy per unit mass, given the kinetic energy per unit mass e_kin, the potential energy per unit mass e_pot, the temperature T, and, optionally, the total specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice.

source

CLIMA.MoistThermodynamics.latent_heat_vapor — Function.

latent_heat_vapor(T)

Return the specific latent heat of vaporization at temperature T.

source

CLIMA.MoistThermodynamics.latent_heat_sublim — Function.

latent_heat_sublim(T)

Return the specific latent heat of sublimation at temperature T.

source

CLIMA.MoistThermodynamics.latent_heat_fusion — Function.

latent_heat_fusion(T)

Return the specific latent heat of fusion at temperature T.

source

CLIMA.MoistThermodynamics.latent_heat_generic — Function.

latent_heat_generic(T, LH_0, cp_diff)

Return the specific latent heat of a generic phase transition between two phases using Kirchhoff's relation.

The latent heat computation assumes constant isobaric specifc heat capacities of the two phases. T is the temperature, LH_0 is the latent heat of the phase transition at T_0, and cp_diff is the difference between the isobaric specific heat capacities (heat capacity in the higher-temperature phase minus that in the lower-temperature phase).

source

CLIMA.MoistThermodynamics.saturation_vapor_pressure — Function.

`saturation_vapor_pressure(T, Liquid())`

Return the saturation vapor pressure over a plane liquid surface at temperature T.

`saturation_vapor_pressure(T, Ice())`

Return the saturation vapor pressure over a plane ice surface at temperature T.

`saturation_vapor_pressure(T, LH_0, cp_diff)`

Compute the saturation vapor pressure over a plane surface by integration of the Clausius-Clepeyron relation.

The Clausius-Clapeyron relation

dlog(p_vs)/dT = [LH_0 + cp_diff * (T-T_0)]/(R_v*T^2)

is integrated from the triple point temperature T_triple, using Kirchhoff's relation

L = LH_0 + cp_diff * (T - T_0)

for the specific latent heat L with constant isobaric specific heats of the phases. The linear dependence of the specific latent heat on temperature T allows analytic integration of the Clausius-Clapeyron relation to obtain the saturation vapor pressure p_vs as a function of the triple point pressure press_triple.

source

CLIMA.MoistThermodynamics.saturation_shum_generic — Function.

saturation_shum_generic(T, ρ[; phase=Liquid()])

Compute the saturation specific humidity over a plane surface of condensate, given the temperature T and the (moist-)air density ρ.

The optional argument phase can be $Liquid()$ or $"ice"$ and indicates the condensed phase.

source

CLIMA.MoistThermodynamics.saturation_shum — Function.

saturation_shum(T, ρ[, q_liq=0, q_ice=0])

Compute the saturation specific humidity, given the temperature T and (moist-)air density ρ.

If the optional liquid, and ice specific humdities q_tot and q_liq are given, the saturation specific humidity is that over a mixture of liquid and ice, computed in a thermodynamically consistent way from the weighted sum of the latent heats of the respective phase transitions (Pressel et al., JAMES, 2015). That is, the saturation vapor pressure and from it the saturation specific humidity are computed from a weighted mean of the latent heats of vaporization and sublimation, with the weights given by the fractions of condensate q_liq/(q_liq + q_ice) and q_ice/(q_liq + q_ice) that are liquid and ice, respectively.

If the condensate specific humidities q_liq and q_ice are not given or are both zero, the saturation specific humidity is that over a mixture of liquid and ice, with the fraction of liquid given by temperature dependent liquid_fraction(T) and the fraction of ice by the complement 1 - liquid_fraction(T).

source

CLIMA.MoistThermodynamics.saturation_shum_from_pressure — Function.

saturation_shum_from_pressure(T, ρ, p_vs)

Compute the saturation specific humidity, given the ambient air temperature T, density ρ, and the saturation vapor pressure p_vs.

source

CLIMA.MoistThermodynamics.liquid_fraction — Function.

liquid_fraction(T[, q_liq=0, q_ice=0])

Return the fraction of condensate that is liquid.

If the optional input arguments q_liq and q_ice are not given or are zero, the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze. If q_liq or q_ice are nonzero, the liquid fraction is computed from them.

source

CLIMA.MoistThermodynamics.heaviside — Function.

heaviside(t)

Return the Heaviside step function at t.

source

Missing docstring.

Missing docstring for phase_partitioning_eq!. Check Documenter's build log for details.

CLIMA.MoistThermodynamics.saturation_adjustment — Function.

saturation_adjustment(e_int, ρ, q_tot[, T_init = T_triple])

Return the temperature that is consistent with the internal energy e_int, (moist-)air density ρ, and total specific humidity q_tot.

The optional input value of the temperature T_init is taken as the initial value of the saturation adjustment iterations.

source

CLIMA.MoistThermodynamics.liquid_ice_pottemp — Function.

liquid_ice_pottemp(T, p[, q_tot=0, q_liq=0, q_ice=0])

Return the liquid-ice potential temperature, given the temperature T, pressure p, total specific humidity q_tot, liquid specific humidity q_liq, and ice specific humidity q_ice.

source