RootSolvers

RootSolvers

CLIMA.RootSolvers — Module.

RootSolvers

Module containing functions for solving roots of non-linear equations. The returned result is a tuple of the root and a Bool indicating convergence.

find_zero(f::F,
           x_0::T,
           x_1::T,
           args::Tuple,
           iter_params::IterParams{R, Int},
           method::RootSolvingMethod
           )::Tuple{T, Bool} where {F, R, T <: Union{R, AbstractArray{R}}}

Interface

find_zero compute x^* such that f(x^*) = 0

Arguments

f equation roots function, where f is callable via f(x, args...)

x_0, x_1 initial guesses

IterParams struct containing absolute tolerance on f(x^*) and maximum iterations

RootSolvingMethod Algorithm to solve roots of the equation:

SecantMethod Secant method

RegulaFalsiMethod Regula Falsi Method

Single example

julia> using RootSolvers
x_0 = 0.0
x_1 = 1.0
f(x, y) = x^2 - y
x_star2 = 10000.0
args = Tuple(x_star2)
x_star = sqrt(x_star2)
tol_abs = 1.0e-3
iter_max = 100

x_root, converged = find_zero(f,
                              x_0,
                              x_1,
                              args,
                              IterParams(tol_abs, iter_max),
                              SecantMethod())

Broadcast example

To broadcast, wrap arguments that need not be broadcasted in a Ref.

julia> using RootSolvers
x_0 = rand(5,5)
x_1 = rand(5,5).+2.0
f(x, y) = x^2 - y
x_star2 = 10000.0
args = Tuple(x_star2)
x_star = sqrt(x_star2)
tol_abs = 1.0e-3
iter_max = 100
x_root, converged = find_zero.(f,
                               x_0,
                               x_1,
                               Ref(args),
                               Ref(IterParams(tol_abs, iter_max)),
                               Ref(SecantMethod()))

source

CLIMA.RootSolvers.find_zero — Function.

find_zero(f, x_0, x_1, args, iter_params, SecantMethod)

Solves the root equation, f, using Secant method. See RootSolvers for more information.

source

find_zero(f, x_0, x_1, args, iter_params, RegulaFalsiMethod)

Solves the root equation, f, using Regula Falsi method. See RootSolvers for more information.

source