Arrays

Arrays

MPIStateArrays

Storage for the state of a discretization.

CLIMA.MPIStateArrays.MPIStateArray — Type.

MPIStateArray{S <: Tuple, T, DeviceArray, N,
              DATN<:AbstractArray{T,N}, Nm1, DAI1} <: AbstractArray{T, N}

N-dimensional MPI-aware array with elements of type T. The dimension N is length(S) + 1. S is a tuple of the first N-1 array dimensions.

Todo

It should be reevaluated whether all this stuff in the type domain is really necessary (some of it was optimistically added for functionality that never panned out)

source

CLIMA.MPIStateArrays.post_Irecvs! — Function.

post_Irecvs!(Q::MPIStateArray)

posts the MPI.Irecv! for Q

source

CLIMA.MPIStateArrays.start_ghost_exchange! — Function.

start_ghost_exchange!(Q::MPIStateArray; dorecvs=true)

Start the MPI exchange of the data stored in Q. If dorecvs is true then post_Irecvs!(Q) is called, otherwise the caller is responsible for this.

This function will fill the send buffer (on the device), copies the data from the device to the host, and then issues the send. Previous sends are waited on to ensure that they are complete.

source

CLIMA.MPIStateArrays.finish_ghost_exchange! — Function.

finish_ghost_exchange!(Q::MPIStateArray)

Complete the exchange of data and fill the data array on the device. Note this completes both the send and the receive communication. For more fine level control see finishghostexchange_recv! and finishghostexchange_send!

source