AtmosDycore

CLIMAAtmosDycore

CLIMA.CLIMAAtmosDycore.getrhsfunction — Function.

getrhsfunction(disc::AbstractAtmosDiscretization)

The spatial discretizations are of the form $Q̇ = f(Q)$, and this function returns the handle to right-hand side function $f$ of the disc

source

CLIMA.CLIMAAtmosDycore.solve! — Function.

solve!(Q, solver::AbstractAtmosODESolver; timeend,
       stopaftertimeend=true, numberofsteps, callbacks)

Solves an ODE using the solver starting from a state Q. The state Q is updated inplace. The final time timeend or numberofsteps must be specified.

A series of optional callback functions can be specified using the tuple callbacks; see GenericCallbacks.

Todo

Currently stopaftertimeend is not used. The idea behind it was that a user might want to stop either one step before or after the final time timeend. This should either be removed or used.

source

Topologies

Topologies encode the connectivity of the elements, spatial domain interval and MPI communication.

Topologies.BrickTopology
Topologies.StackedBrickTopology
Topologies.CubedShellTopology
Topologies.cubedshellmesh
Topologies.cubedshellwarp
Topologies.StackedCubedSphereTopology

Grids

Grids specify the approximation within each element, and any necessary warping.

CLIMA.CLIMAAtmosDycore.Grids.DiscontinuousSpectralElementGrid — Type.

DiscontinuousSpectralElementGrid(topology; FloatType, DeviceArray,
                                 polynomialorder,
                                 meshwarp = (x...)->identity(x))

Generate a discontinuous spectral element (tensor product, Legendre-Gauss-Lobatto) grid/mesh from a topology, where the order of the elements is given by polynomialorder. DeviceArray gives the array type used to store the data (CuArray or Array), and the coordinate points will be of FloatType.

The optional meshwarp function allows the coordinate points to be warped after the mesh is created; the mesh degrees of freedom are orginally assigned using a trilinear blend of the element corner locations.

source

VanillaAtmosDiscretizations

A discretization adds additional information for the atmosphere problem.

CLIMA.CLIMAAtmosDycore.VanillaAtmosDiscretizations.VanillaAtmosDiscretization — Type.

VanillaAtmosDiscretization{nmoist, ntrace}(grid; gravity = true,
viscosity = 0)

Given a 'grid <: AbstractGrid' this construct all the data necessary to run a vanilla discontinuous Galerkin discretization of the the compressible Euler equations with nmoist moisture variables and ntrace tracer variables. If the boolean keyword argument gravity is true then gravity is used otherwise it is not. Isotropic viscosity can be used if viscosity is set to a positive constant.

source

CLIMA.CLIMAAtmosDycore.VanillaAtmosDiscretizations.estimatedt — Function.

estimatedt(disc::VanillaAtmosDiscretization, Q::AtmosStateArray)

Given a discretization disc and a state Q compute an estimate for the time step

Todo

This estimate is currently very conservative, needs to be revisited

source

AtmosStateArray

Storage for the state of a discretization.

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.AtmosStateArray — Type.

AtmosStateArray{S <: Tuple, T, DeviceArray, N,
                DATN<:AbstractArray{T,N}, Nm1, DAI1} <: AbstractArray{T, N}

N-dimensional MPI-aware array with elements of type T. The dimension N is length(S) + 1. S is a tuple of the first N-1 array dimensions.

Todo

It should be reevaluated whether all this stuff in the type domain is really necessary (some of it was optimistically added for functionality that never panned out)

Todo

tag for the MPI message should probably be unified for each AtmosStateArray (right now 888 used is the same for all communication)

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.postrecvs! — Function.

postrecvs!(Q::AtmosStateArray)

posts the MPI.Irecv! for Q

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.startexchange! — Function.

startexchange!(Q::AtmosStateArray; dorecvs=true)

Start the MPI exchange of the data stored in Q. If dorecvs is true then postrecvs!(Q) is called, otherwise the caller is responsible for this.

This function will fill the send buffer (on the device), copies the data from the device to the host, and then issues the send. Previous sends are waited on to ensure that they are complete.

source

CLIMA.CLIMAAtmosDycore.AtmosStateArrays.finishexchange! — Function.

finishexchange!(Q::AtmosStateArray)

Complete the exchange of data and fill the data array on the device

source

LSRKmethods

Methods for time discretization.

CLIMA.CLIMAAtmosDycore.LSRKmethods.LSRK — Type.

LSRK(f, Q; dt, t0 = 0)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

Q̇ = f(Q)

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order, low-storage, Runge–Kutta scheme of Carpenter and Kennedy (1994) (in their notation (5,4) 2N-Storage RK scheme).

References

@TECHREPORT{CarpenterKennedy1994,
  author = {M.~H. Carpenter and C.~A. Kennedy},
  title = {Fourth-order {2N-storage} {Runge-Kutta} schemes},
  institution = {National Aeronautics and Space Administration},
  year = {1994},
  number = {NASA TM-109112},
  address = {Langley Research Center, Hampton, VA},
}

source

CLIMA.CLIMAAtmosDycore.LSRKmethods.updatedt! — Function.

updatedt!(lsrk::LSRK, dt)

Change the time step size to dt for `lsrk.

source