AtmosDycore

AtmosDycore

CLIMA.AtmosDycore.getrhsfunction — Function.

getrhsfunction(disc::AbstractAtmosDiscretization)

The spatial discretizations are of the form $Q̇ = f(Q)$, and this function returns the handle to right-hand side function $f$ of the disc

source

VanillaAtmosDiscretizations

A discretization adds additional information for the atmosphere problem.

CLIMA.AtmosDycore.VanillaAtmosDiscretizations.VanillaAtmosDiscretization — Type.

VanillaAtmosDiscretization{nmoist, ntrace}(grid; gravity = true,
viscosity = 0)

Given a 'grid <: AbstractGrid' this construct all the data necessary to run a vanilla discontinuous Galerkin discretization of the the compressible Euler equations with nmoist moisture variables and ntrace tracer variables. If the boolean keyword argument gravity is true then gravity is used otherwise it is not. Isotropic viscosity can be used if viscosity is set to a positive constant.

source

CLIMA.AtmosDycore.VanillaAtmosDiscretizations.estimatedt — Function.

estimatedt(disc::VanillaAtmosDiscretization, Q::MPIStateArray)

Given a discretization disc and a state Q compute an estimate for the time step

Todo

This estimate is currently very conservative, needs to be revisited

source