DGmethods

ODESolvers

SpaceMethods

Set of abstract types for spatial discretizations

Types

CLIMA.SpaceMethods.AbstractSpaceMethod — Type.

AbstractSpaceMethod

Supertype for spatial discretizations

source

CLIMA.SpaceMethods.AbstractDGMethod — Type.

AbstractDGMethod <: AbstractSpaceMethod

Supertype for discontinuous Galerkin spatial discretizations

source

Functions

CLIMA.SpaceMethods.odefun! — Function.

odefun!(disc::DGBalanceLaw, dQ::MPIStateArray, Q::MPIStateArray, t)

Evaluates the right-hand side of the discontinuous Galerkin semi-discretization defined by disc at time t with state Q. The result is added into dQ. Namely, the semi-discretization is of the form

\[Q̇ = F(Q, t)\]

and after the call dQ += F(Q, t)

source

odefun!(disc::AbstractSpaceMethod, dQ, Q, t)

Evaluates the right-hand side of the spatial discretization defined by disc at time t with state Q. The result is added into dQ. Namely, the semi-discretization is of the form $Q̇ = F(Q, t)$ and after the call dQ += F(Q, t)

Note

There is no generic implementation of this function. This must be implemented for each subtype of AbstractSpaceMethod

source

DGBalanceLawDiscretizations

CLIMA.DGBalanceLawDiscretizations — Module.

DG Balance Law Discretizations module. Attempts to provide a reasonable implementation of a discontinuous Galerkin method (in weak form) on tensor product quadrilateral (2D) and hexahedral (3D) elements for balance laws of the form

\[q_{,t} + Σ_{i=1,...d} F_{i,i} = s\]

where $q$ is the state vector, $F$ is the flux function, and $s$ is the source function. $F$ includes both the "inviscid" and "viscous" fluxes. Note that this is a space only discretization, time must be advanced using some ordinary differential equations methods; see ODESolvers.

Much of the notation used in this module follows Hesthaven and Warburton (2008).

Note

We plan to switch to a skew-symmetric formulation (at which time this note will be removed)

References

@BOOK{HesthavenWarburton2008,
  title = {Nodal Discontinuous {G}alerkin Methods: {A}lgorithms, Analysis,
           and Applications},
  publisher = {Springer-Verlag New York},
  year = {2008},
  author = {Hesthaven, Jan S. and Warburton, Tim},
  volume = {54},
  series = {Texts in Applied Mathematics},
  doi = {10.1007/978-0-387-72067-8}
}

source

Types

CLIMA.DGBalanceLawDiscretizations.DGBalanceLaw — Type.

DGBalanceLaw <: AbstractDGMethod

This contains the necessary information for a discontinuous Galerkin method for balance laws.

See also: Outer constructor DGBalanceLaw

Fields

grid

computational grid / mesh

nstate

number of state

gradstates

tuple of states to take the gradient of

inviscid_flux!

physical inviscid flux function

inviscid_numerical_flux!

inviscid numerical flux function

inviscid_numerical_boundary_flux!

inviscid numerical boundary flux function

Qgrad

storage for the grad

auxstate

auxiliary state array

source!

source function

source

Constructor

CLIMA.DGBalanceLawDiscretizations.DGBalanceLaw — Method.

 DGBalanceLaw(; grid::DiscontinuousSpectralElementGrid, length_state_vector,
              inviscid_flux!, inviscid_numerical_flux!,
              inviscid_numerical_boundary_flux! = nothing,
              auxiliary_state_length=0,
              auxiliary_state_initialization! = nothing,
              source! = nothing)

Constructs a DGBalanceLaw spatial discretization type for the physics defined by inviscid_flux! and source!. The computational domain is defined by grid. The number of state variables is defined by length_state_vector. The user may also specify an auxiliary state which will be unpacked by the compute kernel passed on to the user-defined flux and numerical flux functions. The source function source! is optional.

The inviscid flux function is called with data from a degree of freedom (DOF) as

inviscid_flux!(F, Q, aux, t)

where

F is an MArray of size (dim, length_state_vector) to be filled (note that this is uninitialized so the user must set to zero if is this desired)

Q is the state to evaluate (MArray)

aux is the user-defined auxiliary state (MArray)

t is the current simulation time

Warning: Modifications to Q or aux may cause side effects and should be avoided.

The inviscid numerical flux function is called with data from two DOFs as

inviscid_numerical_flux!(F, nM, QM, auxM, QP, auxP, t)

where

F is an MArray of size (dim, length_state_vector) to be filled with the numerical flux across the face (note that this is uninitialized so user must set to zero if is this desired)

nM is the unit outward normal to the face with respect to the minus side (MVector of length 3)

QM and QP are the minus and plus side states (MArray)

auxM and auxP are the auxiliary states (MArray)

t is the current simulation time

Warning: Modifications to nM, QM, auxM, QP, or auxP may cause side effects and should be avoided.

If grid.topology has a boundary then the function inviscid_numerical_boundary_flux! must be specified. This function is called with the data from the neighbouring DOF as

inviscid_numerical_boundary_flux!(F, nM, QM, auxM, QP, auxP, bctype, t)

where

F is an MArray of size (dim, length_state_vector) to be filled with the numerical flux across the face (note that this is uninitialized so user must set to zero is this desired)

nM is the unit outward normal to the face with respect to the minus side (MVector of length 3)

QM and QP are the minus and plus side states (MArray)

auxM and auxP are the auxiliary states (MArray)

bctype is the boundary condition flag for the connected face and element of grid.elemtobndy

t is the current simulation time

Note: QP and auxP are filled with values based on degrees of freedom referenced in grid.vmapP; QP and auxP may be modified by the calling function.

Warning: Modifications to nM, QM, or auxM may cause side effects and should be avoided.

If present the source function is called with data from a DOF as

source!(S, Q, aux, t)

where S is an MVector of length length_state_vector to be filled; other arguments are the same as inviscid_flux! and the same warning concerning Q and aux applies.

When auxiliary_state_initialization! !== nothing then this is called on the auxiliary state (assuming auxiliary_state_length > 0) as

auxiliary_state_initialization!(aux, x, y, z)

where aux is an MArray to fill with the auxiliary state for a DOF located at Cartesian coordinate locations (x, y, z); see also grad_auxiliary_state! allows the user to take the gradient of a field stored in the auxiliary state.

Note

If (x, y, z), or data derived from this such as spherical coordinates, is needed in the flux or source the user is responsible to storing this in the auxiliary state

Todo

support viscous fluxes (gradstates is in the argument list as part of this future interface)

Revisit how to handle plus side state in inviscid_numerical_boundary_flux! after gradient state is handled (how to or not propagate changes without adverse side effects).

source

Functions

CLIMA.DGBalanceLawDiscretizations.writevtk — Function.

writevtk(prefix, Q::MPIStateArray, disc::DGBalanceLaw [, fieldnames])

Write a vtk file for all the fields in the state array Q using geometry and connectivity information from disc.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames; if not specified the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

source

CLIMA.DGBalanceLawDiscretizations.writevtk_helper — Function.

writevtk_helper(prefix, vgeo::Array, Q::Array, grid, fieldnames)

Internal helper function for writevtk

source

CLIMA.DGBalanceLawDiscretizations.grad_auxiliary_state! — Function.

grad_auxiliary_state!(disc, i, (ix, iy, iz))

Computes the gradient of a the field i of the constant auxiliary state of disc and stores the x, y, z compoment in fields ix, iy, iz of constant auxiliary state.

Note

This only computes the element gradient not a DG gradient. If your constant auxiliary state is discontinuous this may or may not be what you want!

source

Pirated Type Functions

CLIMA.MPIStateArrays.MPIStateArray — Method.

MPIStateArray(disc::DGBalanceLaw; commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

source

CLIMA.MPIStateArrays.MPIStateArray — Method.

MPIStateArray(disc::DGBalanceLaw, initialization!::Function; commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

After allocation the MPIStateArray is initialized using the function initialization! which will be called as:

initialization!(Q, x, y, z, [aux])

where Q is an MArray with the solution state at a single degree of freedom (DOF) to initialize and (x,y,z) is the coordinate point for the allocation. If disc contains an auxiliary data the values of this at the DOF are passed through as an MArray through the aux argument

Note

Q is undef at start the function (i.e., not initialized to zero)

Note

Modifications of the aux array will be discarded.

Todo

GPUify this function to remove host and device data transfers

source

CLIMA.SpaceMethods.odefun! — Method.

odefun!(disc::AbstractSpaceMethod, dQ, Q, t)

Evaluates the right-hand side of the spatial discretization defined by disc at time t with state Q. The result is added into dQ. Namely, the semi-discretization is of the form $Q̇ = F(Q, t)$ and after the call dQ += F(Q, t)

Note

There is no generic implementation of this function. This must be implemented for each subtype of AbstractSpaceMethod

source

Kernels

CLIMA.DGBalanceLawDiscretizations.volumerhs! — Function.

volumerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{ngradstate},
           ::Val{nauxstate}, flux!, source!, rhs::Array, Q, Qgrad, auxstate,
           vgeo, t, D, elems) where {dim, N, nstate, ngradstate,

Computational kernel: Evaluate the volume integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

source

CLIMA.DGBalanceLawDiscretizations.facerhs! — Function.

facerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{ngradstate},
         ::Val{nauxstate}, inviscid_numerical_flux!,
         inviscid_numerical_boundary_flux!, rhs::Array, Q, Qgrad, auxstate,
         vgeo, sgeo, t, vmapM, vmapP, elemtobndy,
         elems) where {dim, N, nstate, ngradstate, nauxstate}

Computational kernel: Evaluate the surface integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

source

CLIMA.DGBalanceLawDiscretizations.initauxstate! — Function.

initauxstate!(::Val{dim}, ::Val{N}, ::Val{nauxstate}, auxstatefun!,
              auxstate, vgeo, elems) where {dim, N, nauxstate}

Computational kernel: Initialize the auxiliary state

See DGBalanceLaw for usage.

source

CLIMA.DGBalanceLawDiscretizations.elem_grad_field! — Function.

elem_grad_field!(::Val{dim}, ::Val{N}, ::Val{nstate}, Q, vgeo, D, elems, s,
                 sx, sy, sz) where {dim, N, nstate}

Computational kernel: Compute the element gradient of state s of Q and store it in sx, sy, and sz of Q.

Warning

This does not compute a DG gradient, but only over the element. If $Q_s$ is discontinuous you may want to consider another approach.

source

DGBalanceLawDiscretizations.NumericalFluxes

CLIMA.DGBalanceLawDiscretizations.NumericalFluxes.rusanov! — Function.

rusanov!(F::MArray, nM, QM, auxM, QP, auxP, t, flux!, wavespeed,
         [preflux = (_...) -> (), correctQ!])

Calculate the Rusanov (aka local Lax-Friedrichs) numerical flux given the plus and minus side states QP and QM using the physical flux function flux! and wavespeed calculation.

The flux! has almost the same calling convention as inviscid_flux! from DGBalanceLaw except that preflux(Q, aux, t) is splatted at the end of the call.

The function wavespeed should return the maximum wavespeed for a state and is called as wavespeed(nM, QM, auxM, t, preflux(QM, auxM, t)...) and wavespeed(nM, QP, auxP, t, preflux(QP, auxP, t)...) where nM is the outward unit normal for the minus side.

When present correctQ!(QM, auxM) and correctQ!(QP, auxP) will be after wavespeed and flux! are called to the user can modify QM and QP before QM - QP is needed; this is useful for correcting Q to include discontinuous reference states.

Todo

We may want to switch to a computed_jump! instead of correctQ! since this would allow the user to better handle round-off error with large background states.

Note

The undocumented arguments PM and PP for the function should not be used by external callers and are used only internally by the function rusanov_boundary_flux!

source