DGmethods_old

DGmethods_old

SpaceMethods

Set of abstract types for spatial discretizations

Types

CLIMA.SpaceMethods.AbstractSpaceMethod — Type.

AbstractSpaceMethod

Supertype for spatial discretizations

source

CLIMA.SpaceMethods.AbstractDGMethod — Type.

AbstractDGMethod <: AbstractSpaceMethod

Supertype for discontinuous Galerkin spatial discretizations

source

Functions

CLIMA.SpaceMethods.odefun! — Function.

odefun!(disc::AbstractSpaceMethod, dQ, Q, t; increment)

Evaluates the right-hand side of the spatial discretization defined by disc at time t with state Q. The result is either added into dQ if increment is true or stored in dQ if it is false. Namely, the semi-discretization is of the form $\dot{Q} = F(Q, t)$ and after the call dQ += F(Q, t) if increment == true or dQ = F(Q, t) if increment == false

Note

There is no generic implementation of this function. This must be implemented for each subtype of AbstractSpaceMethod

source

DGBalanceLawDiscretizations

CLIMA.DGBalanceLawDiscretizations — Module.

DG Balance Law Discretizations module. Attempts to provide a reasonable implementation of a discontinuous Galerkin method (in weak form) on tensor product quadrilateral (2D) and hexahedral (3D) elements for balance laws of the form

\[q_{,t} + Σ_{i=1,...d} F_{i,i} = s\]

where $q$ is the state vector, $F$ is the flux function, and $s$ is the source function. $F$ includes both the "inviscid" and "viscous" fluxes. Note that this is a space only discretization, time must be advanced using some ordinary differential equations methods; see ODESolvers.

The flux function F_{i} is taken to be of the form:

\[F_{i} := F_{i}(q, σ; a)\\ σ = H(q, ∇G(q; a); a)\]

where $a$ is a set of parameters and viscous terms enter through $σ$

The source term is of the form:

\[s := s(q; a)\]

In the code and docs the following terminology is used:

$q$ is referred to as the state

$σ$ is the viscous state

$a$ is the auxiliary state

$F$ is the physical flux

$H$ is the viscous transform

$G$ is the gradient transform

Much of the notation used in this module follows Hesthaven and Warburton (2008).

Note

Currently all the functions take the same parameters and the gradient transform can take a user-specified subset of the state vector.

References

@BOOK{HesthavenWarburton2008,
  title = {Nodal Discontinuous {G}alerkin Methods: {A}lgorithms, Analysis,
           and Applications},
  publisher = {Springer-Verlag New York},
  year = {2008},
  author = {Hesthaven, Jan S. and Warburton, Tim},
  volume = {54},
  series = {Texts in Applied Mathematics},
  doi = {10.1007/978-0-387-72067-8}
}

source

Types

CLIMA.DGBalanceLawDiscretizations.DGBalanceLaw — Type.

DGBalanceLaw <: AbstractDGMethod

This contains the necessary information for a discontinuous Galerkin method for balance laws.

See also: Outer constructor DGBalanceLaw

Fields

grid

computational grid / mesh

nstate

number of state

flux!

physical flux function

numerical_flux!

numerical flux function

numerical_boundary_flux!

numerical boundary flux function

Qvisc

storage for the viscous state

number_gradient_states

number of out states for gradient_transform!

number_viscous_states

number of out states for the viscous_transform!

gradient_transform!

transform from state to variables to take gradient of

viscous_transform!

transform from Q and gradient state to viscous states

viscous_penalty!

penalty for the viscous state computation

viscous_boundary_penalty!

boundary penalty for the viscous state computation (e.g., Dirichlet)

auxstate

auxiliary state array

source!

source function

preodefun!

callback function for before the odefun!

source

Constructor

CLIMA.DGBalanceLawDiscretizations.DGBalanceLaw — Method.

DGBalanceLaw(;grid::DiscontinuousSpectralElementGrid,
             length_state_vector,
             flux!,
             numerical_flux!,
             numerical_boundary_flux! = nothing,
             number_gradient_states = 0,
             number_viscous_states = 0,
             gradient_transform! = nothing,
             viscous_transform! = nothing,
             viscous_penalty! = nothing,
             viscous_boundary_penalty! = nothing,
             auxiliary_state_length = 0,
             auxiliary_state_initialization! = nothing,
             source! = nothing,
             preodefun! = nothing)

Constructs a DGBalanceLaw spatial discretization type for the physics defined by flux! and source!. The computational domain is defined by grid. The number of state variables is defined by length_state_vector. The user may also specify an auxiliary state which will be unpacked by the compute kernel passed on to the user-defined flux and numerical flux functions. The source function source! is optional.

The flux function is called with data from a degree of freedom (DOF) as

flux!(F, Q, V, aux, t)

where

F is an MArray of size (dim, length_state_vector) to be filled (note that this is uninitialized so the user must set to zero if is this desired)

Q is the state to evaluate (MArray)

V is the viscous state to evaluate (MArray)

aux is the user-defined auxiliary state (MArray)

t is the current simulation time

Warning: Modifications to Q or aux may cause side effects and should be avoided.

The numerical flux function is called with data from two DOFs as

numerical_flux!(F, nM, QM, VM, auxM, QP, VP, auxP, t)

where

F is an MVector of length length_state_vector to be filled with the numerical flux across the face (note that this is uninitialized so user must set to zero if is this desired)

nM is the unit outward normal to the face with respect to the minus side (MVector of length 3)

QM and QP are the minus and plus side states (MArray)

VM and VP are the minus and plus viscous side states (MArray)

auxM and auxP are the auxiliary states (MArray)

t is the current simulation time

Warning: Modifications to nM, QM, auxM, QP, or auxP may cause side effects and should be avoided.

If grid.topology has a boundary then the function numerical_boundary_flux! must be specified. This function is called with the data from the neighbouring DOF as

numerical_boundary_flux!(F, nM, QM, VM, auxM, QP, VP, auxP, bctype, t)

where

F is an MArray of size (dim, length_state_vector) to be filled with the numerical flux across the face (note that this is uninitialized so user must set to zero is this desired)

nM is the unit outward normal to the face with respect to the minus side (MVector of length 3)

QM and QP are the minus and plus side states (MArray)

VM and VP are the minus and plus viscous side states (MArray)

auxM and auxP are the auxiliary states (MArray)

bctype is the boundary condition flag for the connected face and element of grid.elemtobndy

t is the current simulation time

Note: QP and auxP are filled with values based on degrees of freedom referenced in grid.vmapP; QP and auxP may be modified by the calling function.

Warning: Modifications to nM, QM, or auxM may cause side effects and should be avoided.

If present the source function is called with data from a DOF as

source!(S, Q, aux, t)

where S is an MVector of length length_state_vector to be filled; other arguments are the same as flux! and the same warning concerning Q and aux applies.

When auxiliary_state_initialization! !== nothing then this is called on the auxiliary state (assuming auxiliary_state_length > 0) as

auxiliary_state_initialization!(aux, x1, x2, x3)

where aux is an MArray to fill with the auxiliary state for a DOF located at Cartesian coordinate locations (x1, x2, x3); see also grad_auxiliary_state! allows the user to take the gradient of a field stored in the auxiliary state.

When viscous terms are needed, the user must specify values for the following keyword arguments:

number_gradient_states (Int)

number_viscous_states (Int)

gradient_transform! (Function)

viscous_transform! (Function)

viscous_penalty! (Function)

viscous_boundary_penalty! (Function); only required if the topology has a boundary

The function gradient_transform! is the implementation of the function G in the module docs; see DGBalanceLawDiscretizations. It is called on each DOF as:

gradient_transform!(G, Q, aux, t)

where G is an MVector of length number_gradient_states to be filled, Q is an MVector containing the states, aux is the full auxiliary state at the DOF, and t is the simulation time.Q

The function viscous_transform! is the implementation of the function H in the module docs; see DGBalanceLawDiscretizations. It transforms the gradient $∇G$ and $q$ into the viscous state $σ$. It is called on each DOF as:

viscous_transform!(V, gradG, Q, aux, t)

where V is an MVector of length number_viscous_states to be filled, gradG is an MMatrix containing the DG-gradient of $G$, Q is an MVector containing the states, aux is the full auxiliary state at the DOF, and t is the simulation time. Note that V is a vector not a matrix so that minimal storage can be used if symmetry can be exploited.

The function viscous_penalty! is the penalty terms to be used for the DG-gradient calculation. It is called with data from two neighbouring degrees of freedom as

viscous_penalty!(V, nM, GM, QM, auxM, GP, QP, auxP, t)

where:

V is an MVector of length number_viscous_states to be filled with the numerical penalty across the face; see below.

nM is the unit outward normal to the face with respect to the minus side (MVector of length 3)

GM and GP are the minus and plus evaluation of gradient_transform! on either side of the face

QM and QP are the minus and plus side states (MArray)

auxM and auxP are the auxiliary states (MArray)

t is the current simulation time

The viscous penalty function should compute on the faces

\[n^{-} \cdot H^{*} - n^{-} \cdot H^{-}\]

where $n^{-} \cdot H^{*}$ is the "numerical-flux" for the viscous state computation and $H^{-}$ is the value of viscous_transform! evaluated on the minus side with $n^{-} \cdot G^{-}$ as an argument.

If grid.topology has a boundary then the function viscous_boundary_penalty! must be specified. This function is called with the data from the neighbouring DOF as

viscous_boundary_penalty!(V, nM, GM, QM, auxM, GP, QP, auxP, bctype, t)

where the required behaviour mimics that of viscous_penalty! and numerical_boundary_flux!.

If preodefun! is called right before the rest of the ODE function, with the main purpose to allow the user to populate/modify the auxiliary state disc.auxstate to be consistent with the current time t and solution vector Q

preodefun!(disc, Q, t)

where disc is the DGBalanceLaw structure and Q is the current state being used to evaluate the ODE function.

notes on `preodefun!`

Unlike the other callbacks, this function is not called at the device (or kernel) level but the host level.

MPI communication of Q occurs after the odefun! and no MPI communication of auxstate is performed (if this is needed we will need to determine a way to handle it in order to overlap communication and computation as well only comm update fields).

Note

If (x1, x2, x3), or data derived from this such as spherical coordinates, is needed in the flux or source the user is responsible to storing this in the auxiliary state

source

Functions

CLIMA.DGBalanceLawDiscretizations.grad_auxiliary_state! — Function.

grad_auxiliary_state!(disc, i, (ix1, ix2, ix3)

Computes the gradient of a the field i of the constant auxiliary state of disc and stores the x1, x2, x3 compoment in fields ix1, ix2, ix3 of constant auxiliary state.

Note

This only computes the element gradient not a DG gradient. If your constant auxiliary state is discontinuous this may or may not be what you want!

source

CLIMA.DGBalanceLawDiscretizations.indefinite_stack_integral! — Function.

indefinite_stack_integral!(disc, f, Q, out_states, [P=disc.auxstate])

Computes an indefinite line integral along the trailing dimension (ξ3 in 3-D and ξ2 in 2-D) up an element stack using state Q

\[∫_{ζ_{0}}^{ζ} f(q; aux, t)\]

and stores the result of the integral in field of P indicated by out_states

The syntax of the integral kernel is:

f(F, Q, aux)

where F is an MVector of length length(out_states), Q and aux are the MVectors for the state and auxiliary state at a single degree of freedom. The function is responsible for filling F.

Requires the isstacked(disc.grid.topology) == true

source

CLIMA.DGBalanceLawDiscretizations.reverse_indefinite_stack_integral! — Function.

reverse_indefinite_stack_integral!(disc, oustate, instate,
                                   [P=disc.auxstate])

reverse previously computed indefinite integral(s) computed with indefinite_stack_integral! to be

\[∫_{ζ}^{ζ_{max}} f(q; aux, t)\]

The states instate[i] is reverse and stored in instate[i].

Requires the isstacked(disc.grid.topology) == true

source

CLIMA.DGBalanceLawDiscretizations.dof_iteration! — Function.

dof_iteration!(dof_fun!::Function, R::MPIStateArray, disc::DGBalanceLaw,
               Q::MPIStateArray)

Iterate over each dof to fill R using the dof_fun!. The syntax of the dof_fun! is

dof_fun!(l_R, l_Q, l_Qvisc, l_aux)

where l_R, l_Q, l_Qvisc, and l_aux are of type MArray filled initially with the values at a single degree of freedom. After the call the values in l_R will be written back to the degree of freedom of R.

source

Pirated Type Functions

CLIMA.MPIStateArrays.MPIStateArray — Method.

MPIStateArray(disc::DGBalanceLaw; nstate=disc.nstate, commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional 'nstate' arguments allows the user to specify a specific number of states. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

source

CLIMA.MPIStateArrays.MPIStateArray — Method.

MPIStateArray(disc::DGBalanceLaw, initialization!::Function; commtag=888)

Given a discretization disc constructs an MPIStateArrays for holding a solution state. The optional commtag allows the user to set the tag to use for communication with this MPIStateArray.

After allocation the MPIStateArray is initialized using the function initialization! which will be called as:

initialization!(Q, x1, x2, x3, aux)

where Q is an MArray with the solution state at a single degree of freedom (DOF) to initialize and (x1, x2, x3) is the coordinate point for the allocation. The auxiliary data the values at the DOF are passed through as an MArray through the aux argument; if disc does not have auxiliary data then the length of the MArray will be zero.

Note

Q is undef at start the function (i.e., not initialized to zero)

Note

Modifications of the aux array will be discarded.

Todo

Remove host and device data transfers.

source

Missing docstring.

Missing docstring for CLIMA.SpaceMethods.odefun!(::DGBalanceLawDiscretizations.DGBalanceLaw, dQ, Q, t). Check Documenter's build log for details.

Kernels

CLIMA.DGBalanceLawDiscretizations.volumerhs! — Function.

volumerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{nviscstate},
           ::Val{nauxstate}, flux!, source!, rhs, Q, Qvisc, auxstate,
           vgeo, t, D, elems) where {dim, N, nstate, nviscstate,

Computational kernel: Evaluate the volume integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

source

CLIMA.DGBalanceLawDiscretizations.facerhs! — Function.

facerhs!(::Val{dim}, ::Val{N}, ::Val{nstate}, ::Val{nviscstate},
         ::Val{nauxstate}, numerical_flux!,
         numerical_boundary_flux!, rhs, Q, Qvisc, auxstate,
         vgeo, sgeo, t, vmapM, vmapP, elemtobndy,
         elems) where {dim, N, nstate, nviscstate, nauxstate}

Computational kernel: Evaluate the surface integrals on right-hand side of a DGBalanceLaw semi-discretization.

See odefun! for usage.

source

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.volumeviscterms!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.faceviscterms!. Check Documenter's build log for details.

CLIMA.DGBalanceLawDiscretizations.initauxstate! — Function.

initauxstate!(::Val{dim}, ::Val{N}, ::Val{nauxstate}, auxstatefun!,
              auxstate, vgeo, elems) where {dim, N, nauxstate}

Computational kernel: Initialize the auxiliary state

See DGBalanceLaw for usage.

source

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.initauxstate!. Check Documenter's build log for details.

CLIMA.DGBalanceLawDiscretizations.elem_grad_field! — Function.

elem_grad_field!(::Val{dim}, ::Val{N}, ::Val{nstate}, Q, vgeo, D, elems, s,
                 sx, sy, sz) where {dim, N, nstate}

Computational kernel: Compute the element gradient of state s of Q and store it in sx, sy, and sz of Q.

Warning

This does not compute a DG gradient, but only over the element. If $Q_s$ is discontinuous you may want to consider another approach.

source

CLIMA.DGBalanceLawDiscretizations.knl_dof_iteration! — Function.

knl_dof_iteration!(::Val{dim}, ::Val{N}, ::Val{nRstate}, ::Val{nstate},
                   ::Val{nviscstate}, ::Val{nauxstate}, dof_fun!, R, Q,
                   QV, auxstate, elems) where {dim, N, nRstate, nstate,
                                               nviscstate, nauxstate}

Computational kernel: fill postprocessing array

See DGBalanceLaw for usage.

source

CLIMA.DGBalanceLawDiscretizations.knl_indefinite_stack_integral! — Function.

knl_indefinite_stack_integral!(::Val{dim}, ::Val{N}, ::Val{nstate},
                                        ::Val{nauxstate}, ::Val{nvertelem},
                                        int_knl!, Q, auxstate, vgeo, Imat,
                                        elems, ::Val{outstate}
                                       ) where {dim, N, nstate, nauxstate,
                                                outstate, nvertelem}

Computational kernel: compute indefinite integral along the vertical stack

See DGBalanceLaw for usage.

source

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.knl_reverse_indefinite_stack_integral!. Check Documenter's build log for details.

DGBalanceLawDiscretizations.NumericalFluxes_old

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.NumericalFluxes_old.rusanov!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DGBalanceLawDiscretizations.NumericalFluxes_old.rusanov_boundary_flux!. Check Documenter's build log for details.