LinearSolvers

LinearSolvers

GeneralizedConjugateResidual

CLIMA.GeneralizedConjugateResidualSolver.GeneralizedConjugateResidual — Type.

GeneralizedConjugateResidual(K, Q, tolerance)

This is an object for solving linear systems using an iterative Krylov method. The constructor parameter K is the number of steps after which the algorithm is restarted (if it has not converged), Q is a reference state used only to allocate the solver internal state, and tolerance specifies the convergence criterion based on the relative residual norm. The amount of memory required by the solver state is roughly (2K + 2) * size(Q). This object is intended to be passed to the linearsolve! command.

This uses the restarted Generalized Conjugate Residual method of Eisenstat (1983).

References

@article{eisenstat1983variational,
  title={Variational iterative methods for nonsymmetric systems of linear equations},
  author={Eisenstat, Stanley C and Elman, Howard C and Schultz, Martin H},
  journal={SIAM Journal on Numerical Analysis},
  volume={20},
  number={2},
  pages={345--357},
  year={1983},
  publisher={SIAM}
}

source

GeneralizedMinimalResidual

CLIMA.GeneralizedMinimalResidualSolver.GeneralizedMinimalResidual — Type.

GeneralizedMinimalResidual(M, Q, tolerance)

This is an object for solving linear systems using an iterative Krylov method. The constructor parameter M is the number of steps after which the algorithm is restarted (if it has not converged), Q is a reference state used only to allocate the solver internal state, and tolerance specifies the convergence criterion based on the relative residual norm. The amount of memory required for the solver state is roughly (M + 1) * size(Q). This object is intended to be passed to the linearsolve! command.

This uses the restarted Generalized Minimal Residual method of Saad and Schultz (1986).

References

@article{saad1986gmres,
  title={GMRES: A generalized minimal residual algorithm for solving nonsymmetric linear systems},
  author={Saad, Youcef and Schultz, Martin H},
  journal={SIAM Journal on scientific and statistical computing},
  volume={7},
  number={3},
  pages={856--869},
  year={1986},
  publisher={SIAM}
}

source

LinearSolvers

CLIMA.LinearSolvers.AbstractLinearSolver — Type.

This is an abstract type representing a generic linear solver.

source

CLIMA.LinearSolvers.AbstractIterativeLinearSolver — Type.

This is an abstract type representing a generic iterative linear solver.

The available concrete implementations are:

GeneralizedConjugateResidual

GeneralizedMinimalResidual

source

CLIMA.LinearSolvers.linearsolve! — Function.

linearsolve!(linearoperator!, Q, Qrhs, solver::AbstractIterativeLinearSolver)

Solves a linear problem defined by the linearoperator! function and the state Qrhs, i.e,

\[ L(Q) = Q_{rhs}\]

using the solver and the initial guess Q. After the call Q contains the solution.

source

CLIMA.LinearSolvers.settolerance! — Function.

settolerance!(solver::AbstractIterativeLinearSolver, tolerance)

Sets the tolerance of the iterative linear solver solver to tolerance.

source