Mesh

Meshing Stuff

Topologies

Topologies encode the connectivity of the elements, spatial domain interval and MPI communication.

Types

CLIMA.Topologies.AbstractTopology — Type.

AbstractTopology{dim}

Represents the connectivity of individual elements, with local dimension dim.

source

CLIMA.Topologies.BoxElementTopology — Type.

BoxElementTopology{dim, T} <: AbstractTopology{dim}

The local topology of a larger MPI-distributed topology, represented by dim-dimensional box elements.

This contains the necessary information for the connectivity elements of the elements on the local process, along with "ghost" elements from neighbouring processes.

Fields

mpicomm

MPI communicator for communicating with neighbouring processes.

elems

Range of element indices

realelems

Range of real (aka nonghost) element indices

ghostelems

Range of ghost element indices

sendelems

Array of send element indices sorted so that

elemtocoord

Element to vertex coordinates; elemtocoord[d,i,e] is the dth coordinate of corner i of element e

Note

currently coordinates always are of size 3 for (x, y, z)

elemtoelem

Element to neighboring element; elemtoelem[f,e] is the number of the element neighboring element e across face f. If there is no neighboring element then elemtoelem[f,e] == e.

elemtoface

Element to neighboring element face; elemtoface[f,e] is the face number of the element neighboring element e across face f. If there is no neighboring element then elemtoface[f,e] == f."

elemtoordr

element to neighboring element order; elemtoordr[f,e] is the ordering number of the element neighboring element e across face f. If there is no neighboring element then elemtoordr[f,e] == 1.

elemtobndy

Element to bounday number; elemtobndy[f,e] is the boundary number of face f of element e. If there is a neighboring element then elemtobndy[f,e] == 0.

nabrtorank

List of the MPI ranks for the neighboring processes

nabrtorecv

Range in ghost elements to receive for each neighbor

nabrtosend

Range in sendelems to send for each neighbor

source

CLIMA.Topologies.BrickTopology — Type.

BrickTopology{dim, T} <: AbstractTopology{dim}

A simple grid-based topolgy. This is a convenience wrapper around BoxElementTopology.

source

CLIMA.Topologies.StackedBrickTopology — Type.

StackedBrickTopology{dim, T} <: AbstractTopology{dim}

A simple grid-based topolgy, where all elements on the trailing dimension are stacked to be contiguous. This is a convenience wrapper around BoxElementTopology.

source

CLIMA.Topologies.CubedShellTopology — Type.

CubedShellTopology{T} <: AbstractTopology{2}

A cube-shell topolgy. This is a convenience wrapper around BoxElementTopology.

source

CLIMA.Topologies.StackedCubedSphereTopology — Type.

StackedCubedSphereTopology{3, T} <: AbstractTopology{3}

A cube-sphere topology. All elements on the same "vertical" dimension are stacked to be contiguous. This is a convenience wrapper around BoxElementTopology.

source

Functions

CLIMA.Topologies.cubedshellmesh — Function.

cubedshellmesh(T, Ne; part=1, numparts=1)

Generate a cubed mesh with each of the "cubes" has an Ne X Ne grid of elements.

The mesh can optionally be partitioned into numparts and this returns partition part. This is a simple Cartesian partition and further partitioning (e.g, based on a space-filling curve) should be done before the mesh is used for computation.

This mesh returns the cubed spehere in a flatten fashion for the vertex values, and a remapping is needed to embed the mesh in a 3-D space.

The mesh structures for the cubes is as follows:

x_2
   ^
   |
4Ne-           +-------+
   |           |       |
   |           |   6   |
   |           |       |
3Ne-           +-------+
   |           |       |
   |           |   5   |
   |           |       |
2Ne-           +-------+
   |           |       |
   |           |   4   |
   |           |       |
 Ne-   +-------+-------+-------+
   |   |       |       |       |
   |   |   1   |   2   |   3   |
   |   |       |       |       |
  0-   +-------+-------+-------+
   |
   +---|-------|-------|------|-> x_1
       0      Ne      2Ne    3Ne

source

CLIMA.Topologies.cubedshellwarp — Function.

cubedshellwarp(a, b, c, R = max(abs(a), abs(b), abs(c)))

Given points (a, b, c) on the surface of a cube, warp the points out to a spherical shell of radius R based on the equiangular gnomonic grid proposed by Ronchi, Iacono, Paolucci (1996) https://dx.doi.org/10.1006/jcph.1996.0047

@article{RonchiIaconoPaolucci1996,
  title={The ``cubed sphere'': a new method for the solution of partial
         differential equations in spherical geometry},
  author={Ronchi, C. and Iacono, R. and Paolucci, P. S.},
  journal={Journal of Computational Physics},
  volume={124},
  number={1},
  pages={93--114},
  year={1996},
  doi={10.1006/jcph.1996.0047}
}

source

Grids

Grids specify the approximation within each element, and any necessary warping.

CLIMA.Grids.DiscontinuousSpectralElementGrid — Type.

DiscontinuousSpectralElementGrid(topology; FloatType, DeviceArray,
                                 polynomialorder,
                                 meshwarp = (x...)->identity(x))

Generate a discontinuous spectral element (tensor product, Legendre-Gauss-Lobatto) grid/mesh from a topology, where the order of the elements is given by polynomialorder. DeviceArray gives the array type used to store the data (CuArray or Array), and the coordinate points will be of FloatType.

The optional meshwarp function allows the coordinate points to be warped after the mesh is created; the mesh degrees of freedom are orginally assigned using a trilinear blend of the element corner locations.

source