ODESolvers

ODESolvers

LowStorageRungeKutta

CLIMA.LowStorageRungeKuttaMethod.LowStorageRungeKutta — Type.

LowStorageRungeKutta(f, Q; dt, t0 = 0)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

Q̇ = f(Q)

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order, low-storage, Runge–Kutta scheme of Carpenter and Kennedy (1994) (in their notation (5,4) 2N-Storage RK scheme).

References

@TECHREPORT{CarpenterKennedy1994,
  author = {M.~H. Carpenter and C.~A. Kennedy},
  title = {Fourth-order {2N-storage} {Runge-Kutta} schemes},
  institution = {National Aeronautics and Space Administration},
  year = {1994},
  number = {NASA TM-109112},
  address = {Langley Research Center, Hampton, VA},
}

source

CLIMA.LowStorageRungeKuttaMethod.updatedt! — Function.

updatedt!(lsrk::LowStorageRungeKutta, dt)

Change the time step size to dt for `lsrk.

source

GenericCallbacks

CLIMA.GenericCallbacks.EveryXWallTimeSeconds — Type.

EveryXWallTimeSeconds(f, time, mpicomm)

This callback will run the function 'f()' every time wallclock time seconds. The mpicomm is used to syncronize runtime across MPI ranks.

source

CLIMA.GenericCallbacks.EveryXSimulationSteps — Type.

EveryXSimulationSteps(f, steps)

This callback will run the function 'f()' every steps of the time stepper

source

ODESolvers

CLIMA.ODESolvers.solve! — Function.

solve!(Q, solver::AbstractODESolver; timeend,
       stopaftertimeend=true, numberofsteps, callbacks)

Solves an ODE using the solver starting from a state Q. The state Q is updated inplace. The final time timeend or numberofsteps must be specified.

A series of optional callback functions can be specified using the tuple callbacks; see GenericCallbacks.

source

CLIMA.ODESolvers.gettime — Function.

gettime(solver::AbstractODESolver)

Returns the current simulation time of the ODE solver solver

source