MoistThermodynamics

MoistThermodynamics Module

The MoistThermodynamics module provides all thermodynamic functions needed for the atmosphere and functions shared across model components. The functions are general for a moist atmosphere that includes suspended cloud condensate in the working fluid; the special case of a dry atmosphere is obtained for zero specific humidities (or simply by omitting the optional specific humidity arguments in the functions that are needed for a dry atmosphere). The general formulation assumes that there are tracers for the total water specific humidity q_tot, the liquid specific humidity q_liq, and the ice specific humidity q_ice to characterize the thermodynamic state and composition of moist air.

There are several types of functions:

Equation of state (ideal gas law):

air_pressure

Specific gas constant and isobaric and isochoric specific heats of moist air:

gas_constant_air

cp_m

cv_m

Specific latent heats of vaporization, fusion, and sublimation:

latent_heat_vapor

latent_heat_fusion

latent_heat_sublim

Saturation vapor pressure and specific humidity over liquid and ice:

sat_vapor_press_liquid

sat_vapor_press_ice

sat_shum

Functions computing energies and inverting them to obtain temperatures

total_energy

internal_energy

air_temperature

Functions to compute temperatures and partitioning of water into phases in thermodynamic equilibrium (when Gibbs' phase rule implies that the entire thermodynamic state of moist air, including the liquid and ice specific humidities, can be calculated from the 3 thermodynamic state variables, such as energy, pressure, and total specific humidity)

liquid_fraction_equil (fraction of condensate that is liquid)

saturation_adjustment (compute temperature from energy, density, and total specific humidity)

Auxiliary functions for diagnostic purposes, e.g., other thermodynamic quantities

liquid_ice_pottemp (liquid-ice potential temperature)

A moist dynamical core that assumes equilibrium thermodynamics can be obtained from a dry dynamical core with total energy as a prognostic variable by including a tracer for the total specific humidity q_tot, using the functions, e.g., for the energies in the module, and computing the temperature T and the liquid and ice specific humidities (q.liq and q.ice) from the internal energy e_int by saturation adjustment:

T = saturation_adjustment(e_int, ρ, q_tot);
q = PhasePartition_equil(T, ρ, q_tot);

here, ρ is the density of the moist air, and the internal energy e_int = e_tot - e_kin - geopotential is the total energy e_tot minus kinetic energy e_kin and potential energy geopotential (all energies per unit mass). No changes to the "right-hand sides" of the dynamical equations are needed for a moist dynamical core that supports clouds, as long as they do not precipitate. Additional source-sink terms arise from precipitation.

Schematically, the workflow in such a core would look as follows:

# initialize
geopotential = grav * z
q_tot          = ...
ρ            = ...

(u, v, w)    = ...
e_kin           = 0.5 * (u^2 + v^2 + w^2)

e_tot        = total_energy(e_kin, geopotential, T, q_tot)

do timestep   # timestepping loop

  # advance dynamical variables by a timestep (temperature typically
  # appears in terms on the rhs, such as radiative transfer)
  advance(u, v, w, ρ, e_tot, q_tot)

  # compute internal energy from dynamic variables
  e_int = e_tot - 0.5 * (u^2 + v^2 + w^2) - geopotential

  # compute temperature, pressure and condensate specific humidities,
  T = saturation_adjustment(e_int, ρ, q_tot);
  q = PhasePartition_equil(T, ρ, q_tot);
  p = air_pressure(T, ρ, q)

end

For a dynamical core that additionally uses the liquid and ice specific humidities q.liq and q.ice as prognostic variables, and thus explicitly allows the presence of non-equilibrium phases such as supercooled water, the saturation adjustment in the above workflow is replaced by a direct calculation of temperature and pressure:

T = air_temperature(e_int, q)
p = air_pressure(T, ρ, q)

Functions

CLIMA.MoistThermodynamics.PhasePartition — Type.

PhasePartition

Represents the mass fractions of the moist air mixture.

Constructors

PhasePartition(q_tot::Real[, q_liq::Real[, q_ice::Real]])
PhasePartition(ts::ThermodynamicState)

See also PhasePartition_equil

Fields

tot

total specific humidity

liq

liquid water specific humidity (default: 0)

ice

ice specific humidity (default: 0)

source

CLIMA.MoistThermodynamics.PhasePartition_equil — Function.

PhasePartition_equil(T, ρ, q_tot)

Partition the phases in equilibrium, returning a PhasePartition object using the liquid_fraction_equil function where

T temperature

ρ (moist-)air density

q_tot total specific humidity

The residual q.tot - q.liq - q.ice is the vapor specific humidity.

source

CLIMA.MoistThermodynamics.ThermodynamicState — Type.

ThermodynamicState{DT}

A thermodynamic state, which can be initialized for various thermodynamic formulations (via its sub-types). All ThermodynamicState's have access to functions to compute all other thermodynamic properties.

source

CLIMA.MoistThermodynamics.PhaseEquil — Type.

PhaseEquil{DT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic equilibrium (therefore, saturation adjustment is needed).

Constructors

PhaseEquil(e_int, q_tot, ρ)

Fields

e_int

internal energy

q_tot

total specific humidity

density

density of air (potentially moist)

T

temperature: computed via saturation_adjustment

source

CLIMA.MoistThermodynamics.PhaseNonEquil — Type.

PhaseNonEquil{DT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic non-equilibrium (therefore, temperature can be computed directly).

Constructors

PhaseNonEquil(e_int, q::PhasePartition, ρ)

Fields

e_int

internal energy

q

phase partition

density

density of air (potentially moist)

source

CLIMA.MoistThermodynamics.LiquidIcePotTempSHumEquil — Function.

LiquidIcePotTempSHumEquil(θ_liq_ice, q_tot, ρ, p)

Constructs a PhaseEquil thermodynamic state from liquid-ice potential temperature.

θ_liq_ice - liquid-ice potential temperature

q_tot - total specific humidity

ρ - density

p - pressure

source

CLIMA.MoistThermodynamics.air_density — Function.

air_density(T, p[, q::PhasePartition])

The (moist-)air density from the equation of state (ideal gas law) where

T air temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

air_density(ts::ThermodynamicState)

The (moist-)air density from the equation of state (ideal gas law), given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.air_pressure — Function.

air_pressure(T, ρ[, q::PhasePartition])

The air pressure from the equation of state (ideal gas law) where:

T air temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

air_pressure(ts::ThermodynamicState)

The air pressure from the equation of state (ideal gas law), given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.air_temperature — Function.

air_temperature(e_int, q::PhasePartition)

The air temperature, where

e_int internal energy per unit mass

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

air_temperature(ts::ThermodynamicState)

The air temperature, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.air_temperature_from_liquid_ice_pottemp — Function.

air_temperature_from_liquid_ice_pottemp(θ_liq_ice, p[, q::PhasePartition])

The air temperature, where

θ_liq_ice liquid-ice potential temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

CLIMA.MoistThermodynamics.cp_m — Function.

cp_m([q::PhasePartition])

The isobaric specific heat capacity of moist air where, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

cp_m(ts::ThermodynamicState)

The isobaric specific heat capacity of moist air, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.cv_m — Function.

cv_m([q::PhasePartition])

The isochoric specific heat capacity of moist air where optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

cv_m(ts::ThermodynamicState)

The isochoric specific heat capacity of moist air given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.dry_pottemp — Function.

dry_pottemp(T, p[, q::PhasePartition])

The dry potential temperature where

T temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

dry_pottemp(ts::ThermodynamicState)

The dry potential temperature, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.exner — Function.

exner(p[, q::PhasePartition])

Compute the Exner function where

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

exner(ts::ThermodynamicState)

Compute the Exner function, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.gas_constant_air — Function.

gas_constant_air([q::PhasePartition])

The specific gas constant of moist air given

q PhasePartition. Without this argument the results are that of dry air.

source

gas_constant_air(ts::ThermodynamicState)

The specific gas constant of moist air given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.Ice — Type.

Ice <: Phase

An ice phase, to dispatch over saturation_vapor_pressure and saturation_shum_generic.

source

CLIMA.MoistThermodynamics.internal_energy — Function.

internal_energy(T[, q::PhasePartition])

The internal energy per unit mass, given a thermodynamic state ts or

T temperature

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

internal_energy(ts::ThermodynamicState)

The internal energy per unit mass, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.internal_energy_sat — Function.

internal_energy_sat(T, ρ, q_tot)

The internal energy per unit mass in thermodynamic equilibrium at saturation where

T temperature

ρ (moist-)air density

q_tot total specific humidity

source

internal_energy_sat(ts::ThermodynamicState)

The internal energy per unit mass in thermodynamic equilibrium at saturation, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.latent_heat_fusion — Function.

latent_heat_fusion(T)

The specific latent heat of fusion where

T temperature

source

latent_heat_fusion(ts::ThermodynamicState)

The specific latent heat of fusion given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.latent_heat_sublim — Function.

latent_heat_sublim(T::Real)

The specific latent heat of sublimation where

T temperature

source

latent_heat_sublim(ts::ThermodynamicState)

The specific latent heat of sublimation given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.latent_heat_vapor — Function.

latent_heat_vapor(T::Real)

The specific latent heat of vaporization where

T temperature

source

latent_heat_vapor(ts::ThermodynamicState)

The specific latent heat of vaporization given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.Liquid — Type.

Liquid <: Phase

A liquid phase, to dispatch over saturation_vapor_pressure and saturation_shum_generic.

source

CLIMA.MoistThermodynamics.liquid_fraction_equil — Function.

liquid_fraction_equil(T[, q::PhasePartition])

The fraction of condensate, assuming phase equilibrium, that is liquid where

T temperature

q PhasePartition

If q.liq or q.ice are nonzero, the liquid fraction is computed from them.

Otherwise, the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze.

source

liquid_fraction_equil(ts::ThermodynamicState)

The fraction of condensate that is liquid given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.liquid_fraction_nonequil — Function.

liquid_fraction_nonequil(T[, q::PhasePartition])

The fraction of condensate, assuming phase non-equilibrium, that is liquid where

T temperature

q PhasePartition

If q.liq or q.ice are nonzero, the liquid fraction is computed from them.

Otherwise, the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze.

Todo

Currently liquid_fraction_nonequil calls liquid_fraction_equil, but we should implement a more general function here.

source

liquid_fraction_nonequil(ts::ThermodynamicState)

The fraction of condensate that is liquid given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.liquid_ice_pottemp — Function.

liquid_ice_pottemp(T, p, q::PhasePartition)

The liquid-ice potential temperature where

T temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

liquid_ice_pottemp(ts::ThermodynamicState)

The liquid-ice potential temperature, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.liquid_ice_pottemp_sat — Function.

liquid_ice_pottemp_sat(T, p[, q::PhasePartition])

The saturated liquid ice potential temperature where

T temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

liquid_ice_pottemp_sat(ts::ThermodynamicState)

The liquid potential temperature given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.moist_gas_constants — Function.

(R_m, cp_m, cv_m, γ_m) = moist_gas_constants([q::PhasePartition])

Wrapper to compute all gas constants at once, where optionally,

q PhasePartition. Without this argument the results are that of dry air.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

Without the specific humidity arguments, the results are that of dry air.

source

(R_m, cp_m, cv_m, γ_m) = moist_gas_constants(ts::ThermodynamicState)

Wrapper to compute all gas constants at once, given a thermodynamic state ts.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

source

CLIMA.MoistThermodynamics.saturation_adjustment — Function.

saturation_adjustment(e_int, ρ, q_tot)

Compute the temperature that is consistent with

e_int internal energy

ρ (moist-)air density

q_tot total specific humidity

See also saturation_adjustment_q_t_θ_l.

source

CLIMA.MoistThermodynamics.saturation_excess — Function.

saturation_excess(T, ρ, q::PhasePartition)

The saturation excess in equilibrium where

T temperature

ρ (moist-)air density

q PhasePartition

The saturation excess is the difference between the total specific humidity q.tot and the saturation specific humidity in equilibrium, and it is defined to be nonzero only if this difference is positive.

source

saturation_excess(ts::ThermodynamicState)

Compute the saturation excess in equilibrium, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.saturation_shum — Function.

saturation_shum(T, ρ[, q::PhasePartition])

Compute the saturation specific humidity, given

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition

If the phase partition q is given, the saturation specific humidity is that over a mixture of liquid and ice, computed in a thermodynamically consistent way from the weighted sum of the latent heats of the respective phase transitions (Pressel et al., JAMES, 2015). That is, the saturation vapor pressure and from it the saturation specific humidity are computed from a weighted mean of the latent heats of vaporization and sublimation, with the weights given by the fractions of condensate q.liq/(q.liq + q.ice) and q.ice/(q.liq + q.ice) that are liquid and ice, respectively.

If the condensate specific humidities q.liq and q.ice are not given or are both zero, the saturation specific humidity is that over a mixture of liquid and ice, with the fraction of liquid given by temperature dependent liquid_fraction_equil(T) and the fraction of ice by the complement 1 - liquid_fraction_equil(T).

source

saturation_shum(ts::ThermodynamicState)

Compute the saturation specific humidity, given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.saturation_shum_generic — Function.

saturation_shum_generic(T, ρ[; phase=Liquid()])

Compute the saturation specific humidity over a plane surface of condensate, given

T temperature

ρ (moist-)air density

and, optionally,

Liquid() indicating condensate is liquid

Ice() indicating condensate is ice

source

CLIMA.MoistThermodynamics.saturation_vapor_pressure — Function.

saturation_vapor_pressure(T, Liquid())

Return the saturation vapor pressure over a plane liquid surface at temperature T.

saturation_vapor_pressure(T, Ice())

Return the saturation vapor pressure over a plane ice surface at temperature T.

saturation_vapor_pressure(T, LH_0, Δcp)

Compute the saturation vapor pressure over a plane surface by integration of the Clausius-Clepeyron relation.

The Clausius-Clapeyron relation

dlog(p_v_sat)/dT = [LH_0 + Δcp * (T-T_0)]/(R_v*T^2)

is integrated from the triple point temperature T_triple, using Kirchhoff's relation

L = LH_0 + Δcp * (T - T_0)

for the specific latent heat L with constant isobaric specific heats of the phases. The linear dependence of the specific latent heat on temperature T allows analytic integration of the Clausius-Clapeyron relation to obtain the saturation vapor pressure p_v_sat as a function of the triple point pressure press_triple.

source

CLIMA.MoistThermodynamics.soundspeed_air — Function.

soundspeed_air(T[, q::PhasePartition])

The speed of sound in air, where

T temperature

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

Without the specific humidity arguments, the results are that of dry air.

source

soundspeed_air(ts::ThermodynamicState)

The speed of sound in air given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.specific_volume — Function.

specific_volume(T, p[, q::PhasePartition])

The (moist-)air specific volume from the equation of state (ideal gas law) where

T air temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

specific_volume(ts::ThermodynamicState)

The (moist-)air specific volume from the equation of state (ideal gas law), given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.total_energy — Function.

total_energy(e_kin, e_pot, T[, q::PhasePartition])

The total energy per unit mass, given

e_kin kinetic energy per unit mass

e_pot potential energy per unit mass

T temperature

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

total_energy(e_kin, e_pot, ts::ThermodynamicState)

The total energy per unit mass given a thermodynamic state ts.

source

CLIMA.MoistThermodynamics.virtual_pottemp — Function.

virtual_pottemp(T, p[, q::PhasePartition])

The virtual temperature where

T temperature

p pressure

and, optionally,

q PhasePartition. Without this argument the results are that of dry air.

source

virtual_pottemp(ts::ThermodynamicState)

The virtual potential temperature, given a thermodynamic state ts.

source